% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability.R
\name{add_f_to_pknca_results}
\alias{add_f_to_pknca_results}
\title{Add bioavailability results to PKNCA results}
\usage{
add_f_to_pknca_results(res_nca, f_aucs)
}
\arguments{
\item{res_nca}{A list containing non-compartmental analysis (NCA) results,
including concentration and dose data.}

\item{f_aucs}{A character vector of the comparing AUC parameter/s including
the prefix f_ (e.g., \code{c("f_aucinf.obs", "f_auclast")}).}
}
\value{
A PKNCA result object which results data frame contains added the
bioavailability calculations requested based on the AUCs provided in \code{f_aucs}.
}
\description{
Add bioavailability results to PKNCA results
}
