% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkconc_ind.R
\name{add_figure_details}
\alias{add_figure_details}
\title{Add Figure Details to Data Frame}
\usage{
add_figure_details(
  adpc,
  title = "",
  subtitle = "",
  collapse_subtitle = "\\n",
  studyid = NULL,
  trt_var,
  plotgroup_vars,
  plotgroup_names,
  xvar_unit,
  xmin = NA,
  xmax = NA,
  footnote = NULL
)
}
\arguments{
\item{adpc}{A data frame containing the data.}

\item{title}{A character string specifying the title for the plot.}

\item{studyid}{A character string specifying the study ID variable.}

\item{trt_var}{A character string specifying the treatment variable.}

\item{plotgroup_vars}{A character vector of the grouping data variables.}

\item{plotgroup_names}{A character vector for the grouping variables names.}

\item{xvar_unit}{A character string for the unit for the x-axis variable.}

\item{xmin}{A numeric value specifying the minimum x-axis limit.}

\item{xmax}{A numeric value specifying the maximum x-axis limit.}

\item{footnote}{A character string specifying plot's manual footnote.}
}
\value{
A data frame with added figure details.
}
\description{
This function adds figure details; title, subtitle, and caption to the data.
}
\author{
Gerardo Rodriguez
}
