% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_lineplot.R
\name{add_optional_layers}
\alias{add_optional_layers}
\title{Helper function to handle optional layers}
\usage{
add_optional_layers(
  plt,
  yaxis_scale,
  show_threshold,
  threshold_value,
  show_dose,
  data,
  time_scale,
  facet_by = NULL
)
}
\arguments{
\item{plt}{The ggplot object to modify}

\item{yaxis_scale}{The scale of the y-axis ("Log" or "Linear")}

\item{show_threshold}{Whether to show a threshold line}

\item{threshold_value}{The value of the threshold line}

\item{show_dose}{Whether to show dose times as vertical lines}

\item{data}{The data used for plotting}

\item{time_scale}{The time scale used for plotting}

\item{facet_by}{Variables to facet the plot by
#' @returns The modified ggplot object with optional layers added}
}
\description{
Helper function to handle optional layers
}
