% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{add_pptx_sl_plot}
\alias{add_pptx_sl_plot}
\title{Add a slide with a plot only}
\usage{
add_pptx_sl_plot(pptx, plot)
}
\arguments{
\item{pptx}{rpptx object}

\item{plot}{ggplot object to show as plot}
}
\value{
rpptx object with slide added
}
\description{
Add a slide with a plot only
}
