% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{add_pptx_sl_plottable}
\alias{add_pptx_sl_plottable}
\title{Add a slide with both a plot and a table}
\usage{
add_pptx_sl_plottable(pptx, df, plot)
}
\arguments{
\item{pptx}{rpptx object}

\item{df}{Data frame to show as table}

\item{plot}{ggplot object to show as plot}
}
\value{
rpptx object with slide added
}
\description{
Add a slide with both a plot and a table
}
