% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{add_pptx_sl_table}
\alias{add_pptx_sl_table}
\title{Add a slide with a table only}
\usage{
add_pptx_sl_table(
  pptx,
  df,
  title = "",
  footer = "Click here for individual results"
)
}
\arguments{
\item{pptx}{rpptx object}

\item{df}{Data frame to show as table}

\item{title}{Title text for the slide}

\item{footer}{Footer text for the slide}
}
\value{
rpptx object with slide added
}
\description{
Add a slide with a table only
}
