% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{add_pptx_sl_title}
\alias{add_pptx_sl_title}
\title{Add a title slide to the rpptx document}
\usage{
add_pptx_sl_title(pptx, title)
}
\arguments{
\item{pptx}{rpptx object}

\item{title}{Title text}
}
\value{
rpptx object with title slide added
}
\description{
Add a title slide to the rpptx document
}
\keyword{internal}
