% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{add_qmd_plot}
\alias{add_qmd_plot}
\title{Helper to create a Quarto code chunk for a plot}
\usage{
add_qmd_plot(plot_expr, use_plotly = FALSE)
}
\arguments{
\item{plot_expr}{Expression for plot.}

\item{use_plotly}{Logical, whether to convert plot to plotly.}
}
\value{
Character vector for Quarto code chunk.
}
\description{
Used internally to format a plot chunk for Quarto documents.
}
