% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{add_qmd_sl_plot}
\alias{add_qmd_sl_plot}
\title{Add a slide to the Quarto document with a single plot}
\usage{
add_qmd_sl_plot(quarto_path, plot, use_plotly = FALSE)
}
\arguments{
\item{quarto_path}{Path to the Quarto (.qmd) file to append to.}

\item{plot}{Expression for plot.}

\item{use_plotly}{Logical, whether to convert plot to plotly.}
}
\value{
Invisibly returns TRUE if the slide was added.
}
\description{
Used internally for dose escalation reporting.
}
