% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{add_qmd_table}
\alias{add_qmd_table}
\title{Helper to create a Quarto code chunk for a table}
\usage{
add_qmd_table(table_expr)
}
\arguments{
\item{table_expr}{Expression for table.}
}
\value{
Character vector for Quarto code chunk.
}
\description{
Used internally to format a table chunk for Quarto documents.
}
