% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_pk_data.R
\name{anonymize_pk_data}
\alias{anonymize_pk_data}
\title{This function anonymizes RO numbers and Study IDs of pk data files.}
\usage{
anonymize_pk_data(path, drug_map = NULL, overwrite = TRUE)
}
\arguments{
\item{path}{path to dataframe in xpt, sas or csv format}

\item{drug_map}{set to NULL}

\item{overwrite}{decides whether file at input location \code{path} is overwritten. default is TRUE}
}
\value{
anonymized dataframe file
}
\description{
This function anonymizes RO numbers and Study IDs of pk data files.
}
\author{
Pascal Bärtschi
}
