% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_operators.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Labels to a dataset}
\usage{
apply_labels(data, labels_df, type)
}
\arguments{
\item{data}{The dataset to which labels will be applied.}

\item{labels_df}{A data frame with three columns: Variable, Label, and Dataset,
for the dataset you are applying it .}

\item{type}{The type variable in labels_df for which the labels are to be applied.}
}
\value{
The same dataset with label attributes applied to all columns.
If a column is not present in the labels list, it will be assigned the name of the col.
}
\description{
This function adds "label" attributes to all columns in a dataset
}
\examples{
\dontrun{
  # Example usage:
  data <- data.frame(USUBJID = c(1, 2, 3), AVAL = c(4, 5, 6))
  labels <- data.frame(
  Variable = c("USUBJID", "AVAL"),
  Label = c("Unique Subject Identifier", "Analysis Value")
  )
  data <- apply_labels(data, labels)
  print(attr(data$A, "label"))
}

}
