% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_selector.R
\name{arrange_plots_by_groups}
\alias{arrange_plots_by_groups}
\title{Arrange Plots by Group Columns}
\usage{
arrange_plots_by_groups(named_list, group_cols)
}
\arguments{
\item{named_list}{A named list of plots, with names in the format 'col1: val1, col2: val2, ...'.}

\item{group_cols}{Character vector of column names to sort by.}
}
\value{
A named list of plots, ordered by the specified group columns.
}
\description{
Orders a named list of plots according to specified grouping columns.
Assumes a specific naming format (i.e, 'col1: val1, col2: val2, ...').
}
