% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bioavailability.R
\name{bioavailability_in_PKNCAresult}
\alias{bioavailability_in_PKNCAresult}
\title{Add bioavailbility to PKNCAresults object}
\usage{
bioavailability_in_PKNCAresult(res_nca, bioavailability)
}
\arguments{
\item{res_nca}{A list containing non-compartmental analysis (NCA) results,
including concentration and dose data.}
}
\description{
This function adds bioavailability (F) data to a PKNCAresults object.
The bioavailabiility is calculated with the calculate_bioavailability function.
}
