% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{bpp_ratios}
\alias{bpp_ratios}
\title{Calculate Blood to Plasma Ratios
This function calculates the blood to plasma ratios for a given data set,
based on the shared time points for each matrix concentration sample.}
\usage{
bpp_ratios(
  data,
  matrix_col,
  conc_col,
  bpp_groups = c("TIME", "USUBJID"),
  blood,
  plasma
)
}
\arguments{
\item{data}{A data frame containing the concentration data.}

\item{matrix_col}{A character string specifying the column name for the matrix type.}

\item{conc_col}{A character string specifying the column name for the concentration data.}

\item{bpp_groups}{A character vector of grouping variables to use for the analysis.
Must include TIME, USUBJID, and optionally, other grouping variables.}

\item{blood}{A character string specifying the value for the blood matrix in matrix_col.}

\item{plasma}{A character string specifying the value for the plasma matrix in matrix_col.}
}
\value{
A data frame containing the blood to plasma ratios.
}
\description{
Calculate Blood to Plasma Ratios
This function calculates the blood to plasma ratios for a given data set,
based on the shared time points for each matrix concentration sample.
}
\examples{
data <- data.frame(
 USUBJID = c("A", "A", "A", "A", "A", "A"),
 TIME = c(0, 1, 2, 0, 1, 2),
 MATRIX = c("BLOOD", "BLOOD", "BLOOD", "PLASMA", "PLASMA", "PLASMA"),
 CONC = c(10, 20, 15, 25, 30, 40)
 )
bpp_ratios(data, "MATRIX", "CONC", bpp_groups = c("TIME", "USUBJID"), "BLOOD", "PLASMA")

}
