% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{calculate_ratio_app}
\alias{calculate_ratio_app}
\title{Links the table ratio of the App with the ratio calculations via PKNCA results}
\usage{
calculate_ratio_app(
  res,
  test_parameter,
  ref_parameter = test_parameter,
  test_group = "(all other levels)",
  ref_group = "PARAM: Analyte01",
  aggregate_subject = "no",
  adjusting_factor = 1,
  custom_pptestcd = NULL
)
}
\arguments{
\item{res}{A PKNCAresult object.}

\item{test_parameter}{Character. The PPTESTCD value to use as test (numerator).}

\item{ref_parameter}{Character. The PPTESTCD value to use as reference (denominator).
Defaults to test_parameter.}

\item{test_group}{Character. The test group (numerator). Default is "(all other levels)".}

\item{ref_group}{Character. The reference group (denominator).}

\item{aggregate_subject}{Character. Aggregation mode: "yes", "no", or "if-needed".}

\item{adjusting_factor}{Numeric that multiplies the calculated ratio. Default is 1.}

\item{custom_pptestcd}{Optional character. If provided, will be used as the PPTESTCD value.}
}
\value{
A data.frame with the calculated ratios for the specified settings.
}
\description{
Links the table ratio of the App with the ratio calculations via PKNCA results
}
