% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_stats.R
\name{calculate_summary_stats}
\alias{calculate_summary_stats}
\title{Calculate Summary Statistics}
\usage{
calculate_summary_stats(resPKNCA = resNCA(), input_groups = "DOSNO")
}
\arguments{
\item{resPKNCA}{A data frame containing results of
Non Compartmental Analysis using PKNCA package}

\item{input_groups}{A character vector specifying the columns to group by.
Here. the hierachrical order matters}
}
\value{
A data frame with summary statistics for each group and parameter.
}
\description{
This function calculates various summary statistics for formatted output of PKNCA::pk.nca().
}
\details{
The function calculates the following statistics for numeric variables:
\itemize{
\item Geometric mean (\code{geomean})
\item Geometric coefficient of variation (\code{geocv})
\item Arithmetic mean (\code{mean})
\item Coefficient of variation (\code{CV})
\item Standard deviation (\code{sd})
\item Minimum value (\code{min})
\item Maximum value (\code{max})
\item Median value (\code{median})
\item Count (\code{count})
}
The function excludes the columns \code{ANALYTE}, \code{DOSEA}, \code{PCSPEC}, \code{STUDYID}, and \code{USUBJID}
from the numeric summarization because they are possible grouping variables
The resulting summary statistics are rounded to three decimal places.
}
\examples{
\dontrun{
data <- data.frame(
  STUDYID = rep(1, 10),
  USUBJID = rep(1:2, each = 5),
  DOSEA = rep(c(10, 20), each = 5),
  ANALYTE = rep(c("A", "B"), each = 5),
  PCSPEC = rep(c("X", "Y"), each = 5),
  VALUE = rnorm(10)
)
input_groups <- c("STUDYID", "USUBJID", "DOSEA", "ANALYTE", "PCSPEC")
calculate_summary_stats(data, input_groups)
}
}
