% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{calculate_table_ratios}
\alias{calculate_table_ratios}
\title{Apply Ratio Calculations to PKNCAresult Object}
\usage{
calculate_table_ratios(res, ratio_table)
}
\arguments{
\item{res}{A PKNCAresult object.}

\item{ratio_table}{Data.frame with columns:
TestParameter, RefParameter, Reference, Test, AggregateSubject,
AdjustingFactor, TestGroups, RefGroups, PPTESTCD.}
}
\value{
The updated PKNCAresult object with added rows in the \code{result} data.frame.
}
\description{
This function takes a PKNCAresult object and a data.frame specifying ratio calculations.
}
