% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{calculate_table_ratios_app}
\alias{calculate_table_ratios_app}
\title{Apply Ratio Calculations to PKNCAresult Object}
\usage{
calculate_table_ratios_app(res, ratio_table)
}
\arguments{
\item{res}{A PKNCAresult object.}

\item{ratio_table}{Data.frame with columns:
Parameter, Reference, Test, AggregateSubject, AdjustingFactor.}
}
\value{
The updated PKNCAresult object with added rows in the \code{result} data.frame.
}
\description{
This function takes a PKNCAresult object and a data.frame specifying ratio calculations
}
