% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_conversions.R
\name{convert_to_iso8601_duration}
\alias{convert_to_iso8601_duration}
\title{Convert Numeric Value and Unit to ISO 8601 Duration}
\usage{
convert_to_iso8601_duration(value, unit)
}
\arguments{
\item{value}{A numeric value representing the time/duration.}

\item{unit}{A character string representing the unit of the time/duration.}
}
\value{
A character string representing the duration in ISO 8601 format.
}
\description{
The function converts a numeric value and its associated time unit into ISO 8601 duration string.
}
\examples{
convert_to_iso8601_duration(200, "h") # Returns "PT200H"
convert_to_iso8601_duration(5, "d")  # Returns "P5D"
}
