% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{create_dose}
\alias{create_dose}
\title{Create PK Dose Dataset}
\usage{
create_dose(ADNCA_conc)
}
\arguments{
\item{ADNCA_conc}{A data frame containing the concentration data.}
}
\value{
A data frame containing the dose data.
}
\description{
This function creates a pharmacokinetic dose dataset from the provided concentration data.
}
\details{
The function performs the following steps:
\itemize{
\item Arranges the data by 'USUBJID' and 'DOSNO'.
\item Groups the data by 'USUBJID' and 'DOSNO'.
\item Selects the first row within each group.
\item Converts 'DOSEA' to numeric and retains the 'ROUTE' as 'IQROUTE'.
}
}
\examples{
\dontrun{
  # Example usage:
  dose_data <- create_dose(conc_data)
}

}
