% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{create_dose_intervals}
\alias{create_dose_intervals}
\title{Create Dose Intervals Dataset}
\usage{
create_dose_intervals(
  mydose = mydose,
  params = c("aucinf.obs", "aucint.last", "auclast", "cmax", "half.life", "tmax",
    "lambda.z", "lambda.z.n.points", "r.squared", "adj.r.squared", "lambda.z.time.first"),
  start_from_last_dose = TRUE
)
}
\arguments{
\item{params}{A character vector specifying the pharmacokinetic parameters to include.}

\item{start_from_last_dose}{Logical defining if start is at time of last dose or C1.}

\item{df_dose}{A PKNCAdose object. Default from the app is \code{mydose}.}
}
\value{
A data frame containing the dose intervals and specified pharmacokinetic parameters.
}
\description{
This function creates a dataset with dose intervals and specified pharmacokinetic parameters.
}
\details{
The function performs the following steps:
\itemize{
\item Creates a vector with all pharmacokinetic parameters.
\item Based on dose times, creates a data frame with start and end times.
\item Adds logical columns for each specified parameter.
}
}
\examples{
\dontrun{
  # Example usage:
  dose_intervals <- create_dose_intervals(df_dose, params)
}

}
