% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_violinboxplot.R
\name{create_filter_expr}
\alias{create_filter_expr}
\title{Helper function create text used to filter data frame}
\usage{
create_filter_expr(boxplotdata, varvalstofilter)
}
\arguments{
\item{boxplotdata}{Data frame to be filtered}

\item{varvalstofilter}{Character vector specifying which variable and value to pre-filter
as \code{colname: value}. By default is NULL (no pre-filtering)}
}
\value{
The filter expression
}
\description{
Helper function create text used to filter data frame
}
