% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_violinboxplot.R
\name{create_filter_text}
\alias{create_filter_text}
\title{Helper function create filter text}
\usage{
create_filter_text(boxplotdata, varvalstofilter)
}
\arguments{
\item{boxplotdata}{Data frame to be filtered}

\item{varvalstofilter}{Character vector specifying which variable and value to pre-filter
as \code{colname: value}. By default is NULL (no pre-filtering)}
}
\value{
A string representing the filter expression
}
\description{
Helper function create filter text
}
