% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{create_html_dose_slides}
\alias{create_html_dose_slides}
\title{Render dose escalation results to HTML via Quarto}
\usage{
create_html_dose_slides(res_dose_slides, path, title)
}
\arguments{
\item{res_dose_slides}{List of results for each dose group.}

\item{path}{Path to the output HTML file.}

\item{title}{Title for the presentation.}
}
\value{
Invisibly returns TRUE if rendering succeeded.
}
\description{
Used internally to create and render a .qmd file to HTML.
}
