% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R
\name{create_indplot}
\alias{create_indplot}
\title{Create an Individual PK Line Plot}
\usage{
create_indplot(
  data,
  selected_usubjids,
  selected_analytes,
  selected_pcspec,
  colorby_var,
  time_scale,
  yaxis_scale,
  cycle,
  facet_by,
  show_threshold,
  threshold_value,
  show_dose,
  palette
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_usubjids, selected_analytes, selected_pcspec, cycle}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to be used for coloring.}

\item{time_scale}{String, either "All Time" or "By Dose Profile".}

\item{yaxis_scale}{String, either "log" or "lin".}

\item{facet_by}{Character vector of column names to facet by.}

\item{show_threshold}{Logical, whether to show threshold line.}

\item{threshold_value}{Numeric, value for the threshold line.}

\item{show_dose}{Logical, whether to show dosing indicators.}

\item{palette}{A named color palette.}

\item{dose_data}{The raw data frame (or a derivative) for calculating dose times.}
}
\value{
A \code{plotly} object.
}
\description{
Combines data processing and plotting for individual concentration-time profiles.
}
