% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mapping.R
\name{create_metabfl}
\alias{create_metabfl}
\title{Create METABFL Column Based on Selected Metabolites}
\usage{
create_metabfl(dataset, metabolites)
}
\arguments{
\item{dataset}{A data frame containing a dataset with a PARAM (parameter) column.}

\item{metabolites}{A character vector of parameter names representing metabolites.}
}
\value{
The input dataset with an additional METABFL column indicating metabolite records ("Y")
or non-metabolite records ("").
}
\description{
Create METABFL Column Based on Selected Metabolites
}
