% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{create_pptx_doc}
\alias{create_pptx_doc}
\title{Create a new PowerPoint document from a template and add a title slide}
\usage{
create_pptx_doc(path, title, template)
}
\arguments{
\item{path}{File path to save the presentation}

\item{title}{Title for the presentation}

\item{template}{Path to PowerPoint template file}
}
\value{
rpptx object
}
\description{
Create a new PowerPoint document from a template and add a title slide
}
