% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_start_impute.R
\name{create_start_impute}
\alias{create_start_impute}
\title{Create C0 Impute Column}
\usage{
create_start_impute(mydata)
}
\arguments{
\item{mydata}{A PKNCAdata object containing concentration and dose data.}
}
\value{
A PKNCAdata object with updated intervals table including start imputation strategies.
}
\description{
Defines an impute column in the intervals of the PKNCAdata object based on data
}
\examples{
\dontrun{
mydata <- create_start_impute(mydata)
}
}
