% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_versioned.R
\name{delete_settings_version}
\alias{delete_settings_version}
\title{Delete a version from a versioned settings list}
\usage{
delete_settings_version(versions, index)
}
\arguments{
\item{versions}{List of version entries.}

\item{index}{Integer index of the version to remove.}
}
\value{
Updated list of version entries.
}
\description{
Removes the version at the given index. Cannot delete the last
remaining version.
}
