% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{.add_pptx_ind_slides}
\alias{.add_pptx_ind_slides}
\title{Add individual-subject slides for one dose group to a pptx object}
\usage{
.add_pptx_ind_slides(pptx, group_data, group_index, in_sections)
}
\arguments{
\item{pptx}{rpptx object}

\item{group_data}{Single element from res_dose_slides}

\item{group_index}{Integer index of the dose group}

\item{in_sections}{Function(id) returning TRUE when id is selected}
}
\value{
List with updated pptx and n_slides count
}
\description{
Add individual-subject slides for one dose group to a pptx object
}
\keyword{internal}
