% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{.add_pptx_main_summary_slide}
\alias{.add_pptx_main_summary_slide}
\title{Add the main summary slide (meanplot / statistics / both) for a dose group}
\usage{
.add_pptx_main_summary_slide(pptx, group_data, i, in_sections)
}
\arguments{
\item{pptx}{An officer pptx object.}

\item{group_data}{One element of res_dose_slides (a dose group).}

\item{i}{Integer index of this dose group.}

\item{in_sections}{Function(id) returning TRUE when the section id is selected.}
}
\value{
Updated pptx object.
}
\description{
Add the main summary slide (meanplot / statistics / both) for a dose group
}
\keyword{internal}
