% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{.add_qmd_group_section_header}
\alias{.add_qmd_group_section_header}
\title{Write a section header slide with the group info table to a qmd file}
\usage{
.add_qmd_group_section_header(quarto_path, res_dose_slides, i, label)
}
\arguments{
\item{quarto_path}{Path to the Quarto (.qmd) file to append to.}

\item{res_dose_slides}{List of results for each dose group.}

\item{i}{Integer index of the dose group.}

\item{label}{Character string used as the # heading text.}
}
\description{
Write a section header slide with the group info table to a qmd file
}
\keyword{internal}
