% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-utils.R
\name{.add_qmd_summary_slides}
\alias{.add_qmd_summary_slides}
\title{Append meanplot/statistics/linplot/boxplot slides for all dose groups to a qmd file}
\usage{
.add_qmd_summary_slides(quarto_path, res_dose_slides, in_sections, use_plotly)
}
\arguments{
\item{quarto_path}{Path to the Quarto (.qmd) file to append to.}

\item{res_dose_slides}{List of results for each dose group.}

\item{in_sections}{Function(id) returning TRUE when the section id is selected.}

\item{use_plotly}{Logical, whether to convert plots to plotly.}
}
\description{
Append meanplot/statistics/linplot/boxplot slides for all dose groups to a qmd file
}
\keyword{internal}
