% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_column_mapping.R
\name{.apply_manual_units}
\alias{.apply_manual_units}
\title{Internal helper that adds unit metadata (AVALU, DOSEU, RRLTU) to the dataset
if defined in the \code{manual_units} list.}
\usage{
.apply_manual_units(dataset, mapping, manual_units)
}
\arguments{
\item{dataset}{A \code{data.frame} or \code{tibble} representing the input data.}

\item{mapping}{A list of selected values from the UI mapping input.}

\item{manual_units}{A named list containing valid unit options.}
}
\value{
A modified \code{dataset} with updated unit columns if applicable.
}
\description{
Internal helper that adds unit metadata (AVALU, DOSEU, RRLTU) to the dataset
if defined in the \code{manual_units} list.
}
