% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compress_range}
\alias{.compress_range}
\title{Compresses a numeric vector into the simplest possible character string that, when evaluated,
will create the same numeric vector.}
\usage{
.compress_range(range_vector)
}
\arguments{
\item{range_vector}{numeric vector with numbers to compress into string}
}
\value{
simplest possible character string representing provided vector
}
\description{
Compresses a numeric vector into the simplest possible character string that, when evaluated,
will create the same numeric vector.
}
