% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.eval_range}
\alias{.eval_range}
\title{Evaluates range notation. If provided notation is invalid, returns NA.}
\usage{
.eval_range(x)
}
\arguments{
\item{x}{character string with range notation, e.g. 1:5.}
}
\value{
numeric vector with specified range of numbers, NA if notation is invalid
}
\description{
Evaluates range notation. If provided notation is invalid, returns NA.
}
