% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_selector.R
\name{.handle_plotly_click}
\alias{.handle_plotly_click}
\title{Update Last Click Data for Slope Selection}
\usage{
.handle_plotly_click(id, last_click_data, manual_slopes, slopes_groups)
}
\arguments{
\item{last_click_data}{A reactiveValues object storing the last clicked data.}

\item{slopes_groups}{A character vector of slope grouping column names.}

\item{dynamic_values}{A named list containing values from the current click event.}

\item{idx_pnt}{The index of the clicked data point.}
}
\value{
Returns \code{NULL} and exits early if data was updated. Otherwise, continues execution.
}
\description{
Checks if the user clicked on a different plot or dataset and updates
\code{last_click_data} accordingly. If an update is needed, the function exits early.
}
