% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotly_empty_plot}
\alias{.plotly_empty_plot}
\title{Generate an Empty Plotly Object}
\usage{
.plotly_empty_plot(message = "No data available")
}
\arguments{
\item{message}{A character string specifying the text to display in the center of the empty plot.
Defaults to \code{"No data available"}.}
}
\value{
A Plotly object representing an empty plot with hidden axes.
}
\description{
This function returns a blank Plotly plot with optional annotation text.
It ensures that when no valid data is available, a meaningful placeholder plot is displayed
instead of causing an error.
}
