% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_column_mapping.R
\name{.rename_columns}
\alias{.rename_columns}
\title{Internal helper to rename dataset columns based on selected column mappings.}
\usage{
.rename_columns(dataset, selected_cols)
}
\arguments{
\item{dataset}{A data frame whose columns are to be renamed.}

\item{selected_cols}{A named list of vectors representing mapped columns}
}
\value{
A data frame with renamed columns based on the provided mappings.
}
\description{
Internal helper to rename dataset columns based on selected column mappings.
}
