% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_selector.R
\name{.update_pknca_with_rules}
\alias{.update_pknca_with_rules}
\title{Apply Slope Rules to Update Data}
\usage{
.update_pknca_with_rules(data, slopes)
}
\arguments{
\item{data}{PKNCA data object}

\item{slopes}{Data frame of slope rules (TYPE, RANGE, REASON, group columns)}
}
\value{
Modified data object with updated flags
}
\description{
Iterates over the given rules and updates the PKNCA object setting inclusion/exclusion flags.
}
