% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{error_plot}
\alias{error_plot}
\title{Create a simple error plot with a message}
\usage{
error_plot(msg)
}
\arguments{
\item{msg}{Character string. The error message to display in the plot.}
}
\value{
A ggplot object with the error message displayed.
}
\description{
This internal function generates a minimal ggplot2 plot displaying a given error message.
It is used to return a plot object with a custom error message.
}
\keyword{internal}
