% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_pkncadata_intervals}
\alias{format_pkncadata_intervals}
\title{Create Dose Intervals Dataset}
\usage{
format_pkncadata_intervals(
  pknca_conc,
  pknca_dose,
  params = c("aucinf.obs", "aucint.last", "auclast", "cmax", "half.life", "tmax",
    "lambda.z", "lambda.z.n.points", "r.squared", "adj.r.squared", "lambda.z.time.first",
    "aucpext.obs", "aucpext.pred", "clast.obs", "cl.obs"),
  start_from_last_dose = TRUE
)
}
\arguments{
\item{pknca_conc}{A PKNCAdose object containing the concentration data.}

\item{pknca_dose}{A PKNCAdose object containing the dose data.}

\item{params}{A character vector specifying the pharmacokinetic parameters to include.}

\item{start_from_last_dose}{Logical defining if start is at time of last dose or C1.}
}
\value{
A data frame containing the dose intervals and specified pharmacokinetic parameters.
}
\description{
This function creates a dataset with dose intervals and specified pharmacokinetic parameters.
}
\details{
The function performs the following steps:
\itemize{
\item Creates a vector with all pharmacokinetic parameters.
\item Based on dose times, creates a data frame with start and end times.
\item Adds logical columns for each specified parameter.
}
}
\examples{
\dontrun{
  # Example usage:
  dose_intervals <- format_pkncadata_intervals(pknca_conc, pknca_dose, params)
}

}
