% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_dose_qc_plot.R
\name{format_unit_string}
\alias{format_unit_string}
\title{Formats a unit string if a unique unit exists}
\usage{
format_unit_string(data, unit_var)
}
\arguments{
\item{data}{The data frame to check.}

\item{unit_var}{The column name of the unit variable.}
}
\value{
A formatted string like " (hr)" or an empty string "".
}
\description{
Formats a unit string if a unique unit exists
}
