% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkcg.R
\name{g_pkcg01_sbs}
\alias{g_pkcg01_sbs}
\title{Wrapper around aNCA::pkcg01() function. Calls the function with \code{SBS} scale argument.}
\usage{
g_pkcg01_sbs(data, ...)
}
\arguments{
\item{data}{Data to be passed into the plotting function.}

\item{...}{Any other parameters to be passed into the plotting function.}
}
\value{
ggplot2 object for pkcg01.
}
\description{
Wrapper around aNCA::pkcg01() function. Calls the function with \code{SBS} scale argument.
}
