% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkconc_ind.R
\name{g_pkconc_ind_lin}
\alias{g_pkconc_ind_lin}
\title{Wrapper around aNCA::pkcg01() function. Calls the function with \code{LIN} scale argument.}
\usage{
g_pkconc_ind_lin(data, ...)
}
\arguments{
\item{data}{Data to be passed into the plotting function.}

\item{...}{Any other parameters to be passed into the plotting function.}
}
\value{
ggplot2 object for pkcg01.
}
\description{
Wrapper around aNCA::pkcg01() function. Calls the function with \code{LIN} scale argument.
}
