% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkconc.R
\name{g_pkconc_log}
\alias{g_pkconc_log}
\title{Wrapper around aNCA::pkcg02() function. Calls the function with \code{LOG} scale argument.}
\usage{
g_pkconc_log(data, ...)
}
\arguments{
\item{data}{Data to be passed into the plotting function.}

\item{...}{Any other parameters to be passed into the plotting function.}
}
\value{
ggplot2 object for pkcg02.
}
\description{
Wrapper around aNCA::pkcg02() function. Calls the function with \code{LOG} scale argument.
}
