% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coldefs.R
\name{generate_col_defs}
\alias{generate_col_defs}
\title{Generate Column Definitions for Tables}
\usage{
generate_col_defs(data)
}
\arguments{
\item{data}{A data frame containing the data to be displayed in the reactable table.}
}
\value{
A named list of column definitions for the reactable table.
}
\description{
This function generates column definitions for a reactable table,
including both the column heading and the label (if available).
}
\examples{
\dontrun{
  data <- data.frame(
    USUBJID = c(1, 2, 3),
    AVAL = c(4, 5, 6)
  )
  attr(data$USUBJID, "label") <- "Unique Subject Identifier"
  attr(data$AVAL, "label") <- "Analysis Value"
  col_defs <- generate_col_defs(data)
}

}
