% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_operators.R
\name{generate_tooltip_text}
\alias{generate_tooltip_text}
\title{Generate HTML Tooltip Text}
\usage{
generate_tooltip_text(data, labels_df, tooltip_vars, type)
}
\arguments{
\item{data}{A data.frame with the source data.}

\item{labels_df}{A data.frame used by \code{get_label()} to find variable labels.}

\item{tooltip_vars}{A character vector of column names to include in the tooltip.}

\item{type}{A character string specifying the label type for \code{get_label()}.}
}
\value{
A character vector of formatted HTML tooltip strings.
}
\description{
Creates a character vector of HTML tooltips for each row of a data frame,
suitable for use with \code{ggplotly}.
}
\examples{
\dontrun{
# Dummy get_label function for demonstration purposes
get_label <- function(labels, var, type) {
  # A real function would look up the label in a data frame.
  # This dummy function just capitalizes the variable name.
  toupper(var)
  }

# Sample data
my_data <- data.frame(
  USUBJID = c("Subject-01", "Subject-02"),
  DOSE = c(100, 200),
  RESPONSE = c(5.4, 8.1)
  )

my_labels <- data.frame() # Dummy labels object

vars_to_show <- c("USUBJID", "DOSE", "RESPONSE")

# Generate the tooltip text vector
tooltips <- generate_tooltip_text(my_data, my_labels, vars_to_show, "ADPC")

# View the output
tooltips
# [1] "<b>USUBJID</b>: Subject-01<br><b>DOSE</b>: 100<br><b>RESPONSE</b>: 5.4"
# [2] "<b>USUBJID</b>: Subject-02<br><b>DOSE</b>: 200<br><b>RESPONSE</b>: 8.1"

# Example of adding it to the data frame for plotting
my_data$tooltip <- tooltips
}

}
