% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_meanplot.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Helper Function: Calculate the Geometric Mean}
\usage{
geometric.mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether NA values should be removed.}
}
\value{
The geometric mean of the input vector.
}
\description{
Helper Function: Calculate the Geometric Mean
}
