% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_conversions.R
\name{get_conversion_factor}
\alias{get_conversion_factor}
\title{Transform Units}
\usage{
get_conversion_factor(initial_unit, target_unit)
}
\arguments{
\item{initial_unit}{A character string representing the initial unit.}

\item{target_unit}{A character string representing the target unit.}
}
\value{
A numeric value for the conversion factor from the initial to the target unit,
or NA if the units are not convertible.
}
\description{
This function transforms a value from an initial unit to a target unit.
}
\examples{
get_conversion_factor("meter", "kilometer")
get_conversion_factor("sec", "min")
}
