% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_operators.R
\name{get_label}
\alias{get_label}
\title{Get the Label of a Heading}
\usage{
get_label(variable, type)
}
\arguments{
\item{variable}{The variable for which the label is to be retrieved.}

\item{type}{The type of the dataset for which the label is to be retrieved.}
}
\value{
The label of the heading if it exists in the labels file,
otherwise "No label available".
}
\description{
This function retrieves the label of a heading from a labels file.
}
\examples{
\dontrun{
 # Example usage:
 LABELS <- data.frame(
 Variable = c("USUBJID", "AVAL"),
 Label = c("Unique Subject Identifier", "Analysis Value"),
 Dataset = c("ADPC", "ADPC")
 )
 get_label("USUBJID", "ADPC")  # Returns "Unique Subject Identifier"
 get_label("AGE", "ADPC")  # Returns "No label available"
 }

}
