% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_selector.R
\name{handle_hl_adj_change}
\alias{handle_hl_adj_change}
\title{Handle half-life adjustment changes}
\usage{
handle_hl_adj_change(new_pknca_data, old_pknca_data, plot_outputs)
}
\arguments{
\item{new_pknca_data}{New PKNCA data object}

\item{old_pknca_data}{Previous PKNCA data object}

\item{plot_outputs}{Current plot outputs (named list)}
}
\value{
Updated plot_outputs (named list)
}
\description{
Updates only the plots affected by changes in half-life inclusion/exclusion columns.
}
