% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{handle_imputation}
\alias{handle_imputation}
\title{Handle Imputation for PKNCA Data}
\usage{
handle_imputation(data)
}
\arguments{
\item{data}{A PKNCAdata object.}
}
\value{
A PKNCAdata object with imputation rules applied.
}
\description{
Applies imputation rules to a PKNCAdata object, imputing start values and then
selectively removing imputation for parameters that are not dependent on AUC.
}
