% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_violinboxplot.R
\name{handle_tooltips}
\alias{handle_tooltips}
\title{Tooltip construction helper}
\usage{
handle_tooltips(box_data, tooltip_vars, labels_df)
}
\arguments{
\item{box_data}{the boxplot dataframe}

\item{tooltip_vars}{character vector of tooltip variables to extract}

\item{labels_df}{data.frame used for label lookups in tooltips}
}
\value{
box_data with added tooltip_text column
}
\description{
Tooltip construction helper
}
