% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkcg.R
\name{keep_blq_timepoints}
\alias{keep_blq_timepoints}
\title{Helper function for BLQ processing}
\usage{
keep_blq_timepoints(plot_data, xvar, mean_group_var)
}
\arguments{
\item{plot_data}{Data frame containing the data to be processed.}

\item{xvar}{Character string of the variable name for the x-axis.}

\item{mean_group_var}{Character string specifying the grouping variable to plot by group.}
}
\value{
A data frame with valid timepoints based on BLQ criteria.
}
\description{
Helper function for BLQ processing
}
