% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_pkcl01.R
\name{l_pkcl01_tad}
\alias{l_pkcl01_tad}
\title{Wrapper around aNCA::l_pkcl01() for TAD-based concentration listings.}
\usage{
l_pkcl01_tad(data, ...)
}
\arguments{
\item{data}{Data to be passed into the listing function.}

\item{...}{Any other parameters to be passed into the listing function.}
}
\value{
A named list of listing_df objects.
}
\description{
Wrapper around aNCA::l_pkcl01() for TAD-based concentration listings.
}
