% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_pknca_cdisc}
\alias{dict_pknca_cdisc}
\title{dict_pknca_cdisc}
\format{
A data frame with 120 rows and 6 variables:
\describe{
\item{PKNCA}{PKNCA term}
\item{PPTESTCD}{CDISC term}
\item{Label}{Description of the term}
\item{PPTEST}{Official CDISC term}
\item{PPTESTCD}{Official CDISC short code term}
\item{unit_type}{Type of unit associated with the term}
\item{pretty_name}{PKNCA pretty name of the term}
\item{description}{PKNCA description of the term}
\item{FUN}{PKNCA function used to calculate the parameter}
\item{is_cdisc_sure}{Logical indicating if the term is a CDISC official name}\
}
}
\source{
Generated for use in the \code{translate_nomenclature} function.
}
\usage{
dict_pknca_cdisc
}
\description{
A dataset containing the mapping between PKNCA terms and CDISC terms. It mainly involves:
\itemize{
\item Parameters: PPTEST, PPTESTCD
}
}
\keyword{datasets}
