% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{log_conversion}
\alias{log_conversion}
\title{Log conversions applied to dataset}
\usage{
log_conversion(row, vol, volu, u_vol_new, denom_unit, concu, verbose = TRUE)
}
\arguments{
\item{row}{The row number where the conversion is applied.}

\item{vol}{The original volume value.}

\item{volu}{The original volume unit.}

\item{u_vol_new}{The new volume value after conversion.}

\item{denom_unit}{The denominator unit derived from the concentration unit.}

\item{concu}{The concentration unit.}

\item{verbose}{A logical indicating whether to log the conversion (default: TRUE).}
}
\value{
NULL if units remained the same, or log info of the conversions that were applied
}
\description{
This helper function logs the conversion of volume units in a dataset.
}
\keyword{internal}
