% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_nca_parameters}
\alias{metadata_nca_parameters}
\title{metadata_nca_parameters}
\format{
A data frame with 123 rows and 6 variables:
\describe{
\item{PKNCA}{PKNCA term}
\item{PPTESTCD}{CDISC term}
\item{PPTEST}{Official CDISC term}
\item{input_names}{Combination of PPTESTCD + ": " + PPTEST. Used for App inputs}
\item{FUN}{PKNCA function used to calculate the parameter}
\item{description}{PKNCA description of the term}
\item{is_cdisc_sure}{Logical indicating if the term is a CDISC official name}
\item{unit_type}{Type of unit associated with the term}
\item{TYPE}{Type of data associated with the parameter/term}
\item{CAT}{Arbitrary assigned subclass for the parameter/term}
\item{Depends}{PKNCA derived. Designates all directly used parameters in calculation}
\item{can_excretion}{Logical. Indicates if the parameter can be used in excretion analysis}
\item{can_non_excretion}{Logical.
Indicates if the parameter can be used in non-excretion analysis}
\item{can_single_dose}{Logical.
Indicates if the parameter can be used in single dose analysis}
\item{can_multiple_dose}{Logical.
Indicates if the parameter can be used in multiple dose analysis}
\item{can_extravascular}{Logical.
Indicates if the parameter can be used in extravascular analysis}
}
}
\source{
Generated for use in the \code{translate_nomenclature} function.
}
\usage{
metadata_nca_parameters
}
\description{
A dataset containing the mapping between PKNCA terms and CDISC terms. It mainly involves:
\itemize{
\item Parameters: PPTEST, PPTESTCD
}
}
\keyword{datasets}
