% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkcg.R
\name{p_pkcg03_log_dose}
\alias{p_pkcg03_log_dose}
\title{Wrapper around aNCA::pkcg03() function. Mean log plot grouped by dose.}
\usage{
p_pkcg03_log_dose(data, ...)
}
\arguments{
\item{data}{Data to be passed into the plotting function.}

\item{...}{Any other parameters to be passed into the plotting function.}
}
\value{
ggplot2 object for pkcg03.
}
\description{
Wrapper around aNCA::pkcg03() function. Mean log plot grouped by dose.
}
