% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tlg.R
\name{parse_tlg_definitions}
\alias{parse_tlg_definitions}
\title{Parses the tlg definitions described in yaml file into a list object.
Custom YAML parsing is needed for templating - the user can define templates for a definition.
The function detects appropriate keywords and copies information from specified templates.}
\usage{
parse_tlg_definitions(path = NULL, defs = NULL)
}
\arguments{
\item{path}{Path to a YAML file containing TLG definitions.}

\item{defs}{A list of TLG definitions read from a YAML file.}
}
\value{
A full list of available TLG definitions.
}
\description{
Parses the tlg definitions described in yaml file into a list object.
Custom YAML parsing is needed for templating - the user can define templates for a definition.
The function detects appropriate keywords and copies information from specified templates.
}
\details{
If both \code{path} and \code{defs} are provided, \code{path} will be used and definitions will be loaded from
the file. If neither is given, as a default, the function will load system file in
\code{shiny/tlg.yaml}.
}
