% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_partialAUCInput.R
\name{partial_auc_input}
\alias{partial_auc_input}
\title{Define the UI for a single partial AUC input}
\usage{
partial_auc_input(id, min_sel_value = 0, max_sel_value = NULL)
}
\arguments{
\item{id}{A unique identifier for the input.}

\item{min_sel_value}{The default minimum value for the input.}

\item{max_sel_value}{The default maximum value for the input.}
}
\value{
A Shiny UI component for creating a partial AUC input.
}
\description{
Define the UI for a single partial AUC input
}
