% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_pknca_results.R
\name{pivot_wider_pknca_results}
\alias{pivot_wider_pknca_results}
\title{Reshape PKNCA Results}
\usage{
pivot_wider_pknca_results(myres)
}
\arguments{
\item{myres}{The output of PKNCA::pk.nca}
}
\value{
A data frame which provides an easy overview on the results from the NCA
in each profile/subject and how it was computed lambda (half life) and the results
of the NCA parameters (cmax, AUC, AUClast)
}
\description{
This function reshapes the structure of the results produced by the main function
of the PKNCA package (pk.nca) in a way that each row represents all the main results
summarized for each profile in each individual/patient. Excluding the ID variables,
each column name corresponds with a calculated parameter and between brackets its
corresponding units. AUC intervals, if present, are be added as additional columns.
}
