% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_extra_parameters.R
\name{pk.calc.ermax}
\alias{pk.calc.ermax}
\title{Calculate the maximum excretion rate}
\usage{
pk.calc.ermax(conc, volume, time, duration.conc, check = TRUE)
}
\arguments{
\item{conc}{The concentration in the excreta (e.g., urine or feces)}

\item{volume}{The volume (or mass) of the sample}

\item{time}{The starting time of the collection interval}

\item{duration.conc}{The duration of the collection interval}

\item{check}{Should the concentration data be checked?}
}
\value{
The maximum excretion rate, or NA if not available
}
\description{
Calculate the maximum excretion rate
}
