% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_extra_parameters.R
\name{pk.calc.ertlst}
\alias{pk.calc.ertlst}
\title{Calculate the midpoint collection time of the last measurable excretion rate}
\usage{
pk.calc.ertlst(conc, volume, time, duration.conc, check = TRUE)
}
\arguments{
\item{conc}{The concentration in the excreta (e.g., urine or feces)}

\item{volume}{The volume (or mass) of the sample}

\item{time}{The starting time of the collection interval}

\item{duration.conc}{The duration of the collection interval}

\item{check}{Should the concentration and time data be checked?}
}
\value{
The midpoint collection time of the last measurable excretion rate, or NA/0 if not available
}
\description{
Calculate the midpoint collection time of the last measurable excretion rate
}
