% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_extra_parameters.R
\name{pk.calc.ertmax}
\alias{pk.calc.ertmax}
\title{Calculate the midpoint collection time of the maximum excretion rate}
\usage{
pk.calc.ertmax(
  conc,
  volume,
  time,
  duration.conc,
  check = TRUE,
  first.tmax = NULL,
  options = list()
)
}
\arguments{
\item{conc}{The concentration in the excreta (e.g., urine or feces)}

\item{volume}{The volume (or mass) of the sample}

\item{time}{The starting time of the collection interval}

\item{duration.conc}{The duration of the collection interval}

\item{check}{Should the concentration and time data be checked?}

\item{first.tmax}{If TRUE, return the first time of maximum excretion rate; otherwise, return the last}

\item{options}{List of PKNCA global options set}
}
\value{
The midpoint collection time of the maximum excretion rate, or NA if not available
}
\description{
Calculate the midpoint collection time of the maximum excretion rate
}
