% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_extra_parameters.R
\name{pk.calc.volpk}
\alias{pk.calc.volpk}
\title{Calculate the total urine volume}
\usage{
pk.calc.volpk(volume)
}
\arguments{
\item{volume}{The volume (or mass) of the sample}
}
\value{
The sum of urine volumes for the interval
}
\description{
Calculate the total urine volume
}
