% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkcg.R
\name{pkcg03}
\alias{pkcg03}
\title{Generate PK Concentration-Time Profile Plots}
\usage{
pkcg03(
  adpc = data(),
  xvar = "NFRLT",
  yvar = "AVAL",
  xvar_unit = "RRLTU",
  yvar_unit = "AVALU",
  xbreaks_var = "NFRLT",
  xbreaks_mindist = 0.5,
  xmin = NA,
  xmax = NA,
  ymin = NA,
  ymax = NA,
  xlab = paste0("!", xvar, " [$", xvar_unit, "]"),
  ylab = paste0("!", yvar, " [$", yvar_unit, "]"),
  title = NULL,
  subtitle = NULL,
  footnote = NULL,
  plotgroup_vars = c("ROUTE", "PCSPEC", "PARAM"),
  plotgroup_names = list(ROUTE = "Route", PCSPEC = "Specimen", PARAM = "Analyte"),
  scale = c("LIN", "LOG", "SBS")[1],
  studyid = "STUDYID",
  mean_group_var = "DOSEA",
  plotly = TRUE,
  summary_method = "Mean_sdi",
  whiskers_lwr_upr = c("Both", "Upper", "Lower")[1]
)
}
\arguments{
\item{adpc}{A data frame containing the data.}

\item{xvar}{A character string of the variable name for the x-axis.}

\item{yvar}{A character string of the variable name for the y-axis.}

\item{xvar_unit}{A character string of the unit for the x-axis variable.}

\item{yvar_unit}{A character string of the unit for the y-axis variable.}

\item{xbreaks_var}{A character string of the x-axis breaks.}

\item{xbreaks_mindist}{A numeric value for minimum distance between x-axis breaks in centimeters.}

\item{xmin}{A numeric value specifying the minimum x-axis limit.}

\item{xmax}{A numeric value specifying the maximum x-axis limit.}

\item{ymin}{A numeric value for the minimum y-axis limit.}

\item{ymax}{A numeric value for the maximum y-axis limit.}

\item{xlab}{Character for x-axis label. Defaults: \code{xvar} label & \code{xvar_unit}.}

\item{ylab}{Character for y-axis label. Defaults: \code{yvar} label & \code{yvar_unit}.}

\item{title}{A character string for plot title.}

\item{subtitle}{A character string for plot subtitle.}

\item{footnote}{A character string of a manual footnote for the plot.}

\item{plotgroup_vars}{A character vector of the variables to group data.}

\item{plotgroup_names}{A character vector of the grouping variable names.}

\item{scale}{A character string:"LIN", "LOG", or "SBS", defining the y-axis scale.}

\item{studyid}{A character string specifying the study ID variable.}

\item{mean_group_var}{A character string specifying the grouping variable to plot by group.}

\item{plotly}{Logical indicating whether to return plotly objects. Defaults to TRUE.}

\item{summary_method}{A character string specifying the stat method to summarize observations.}

\item{whiskers_lwr_upr}{A chararcter string specifying the whisker type (upper, lower or both)}

\item{color_var}{A character string of the variable name for the color.}

\item{color_var_label}{A character string of the color label.}
}
\value{
A list of ggplot objects for each unique group.
}
\description{
This function generates a list of ggplots for PK concentration-time profiles.
}
\examples{
\dontrun{
  adpc <- read.csv("inst/shiny/data/DummyRO_ADNCA.csv")
  attr(adpc[["AFRLT"]], "label") <- "Actual time from first dose"
  attr(adpc[["AVAL"]], "label") <- "Analysis val"

  plots_lin <- pckg03(adpc = adpc, xmax = 1)
  plots_log <- pckg03(adpc = adpc, color_var = "USUBJID", scale = "LOG")
  plots_sbs <- pckg03(
    adpc = adpc,
    color_var = "USUBJID",
    xbreaks_var = "NFRLT",
    xmin = 100,
    xmax = 1000,
    scale = "SBS",
    plotly = TRUE,
    whiskers_lwr_upr = "Both"
  )
}
}
\author{
Kezia Kobana
magic numbers for footnote position and margin, work in app up to 4 lines
}
