% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.R
\name{PKNCA_hl_rules_exclusion}
\alias{PKNCA_hl_rules_exclusion}
\title{Exclude NCA results based on user-defined rules over the half-life related parameters
This function applies exclusion rules to the NCA results based on user-defined parameters.}
\usage{
PKNCA_hl_rules_exclusion(res, rules)
}
\arguments{
\item{res}{A PKNCAresults object containing the NCA results.}

\item{rules}{A list of exclusion rules where each rule is a named vector.}
}
\value{
A PKNCAresults object with the exclusions applied.
}
\description{
Exclude NCA results based on user-defined rules over the half-life related parameters
This function applies exclusion rules to the NCA results based on user-defined parameters.
}
\details{
The function iterates over the rules and applies the exclusion criteria to the NCA results.
For any parameter that is not aucpext.obs or aucpext.pred it applies a minimum threshold,
and for aucpext.obs and aucpext.pred it applies a maximum threshold.
}
