% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_impute_method_additions.R
\name{PKNCA_impute_method_start_log}
\alias{PKNCA_impute_method_start_log}
\title{This function imputes the start concentration using the log slope method.}
\usage{
PKNCA_impute_method_start_log(conc, time, start, end, ..., options = list())
}
\arguments{
\item{conc}{Numeric vector of concentrations.}

\item{time}{Numeric vector of times corresponding to the concentrations.}

\item{start}{Numeric value indicating the start/dose time.}

\item{end}{Numeric value indicating the end time.}

\item{...}{Additional arguments (currently not used).}

\item{options}{List of options (currently not used).}
}
\value{
A data frame with imputed start concentration.
}
\description{
This function imputes the start concentration using the log slope method.
}
\examples{
conc <- c(1, 2, 3, 4, 5)
time <- c(0, 1, 2, 3, 4)
start <- 0
end <- 4
PKNCA_impute_method_start_log(conc, time, start, end)
}
