% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.R
\name{PKNCA_left_join_conc_dose}
\alias{PKNCA_left_join_conc_dose}
\title{Full Join of Concentration and Dose Data for PKNCA}
\usage{
PKNCA_left_join_conc_dose(o_conc, o_dose)
}
\arguments{
\item{o_conc}{A PKNCA concentration object (PKNCAconc).}

\item{o_dose}{A PKNCA dose object (PKNCAdose).}
}
\value{
A data frame containing the concentration data with the corresponding dose information.
}
\description{
This function combines concentration and dose PKNCA objects into a single data frame.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts necessary columns from the concentration and dose objects.
\item Combines the data using \code{bind_rows}, adding a helper column to distinguish dose rows.
\item Fills values in the combined data.
\item Filters out dose rows and returns the final combined data frame.
s
}
}
\examples{
# Assuming `o_conc` and `o_dose` are valid PKNCA objects and present common groups:
d_conc <- data.frame(
  subj = 1,
  analyte = rep(c("A", "B"), each = 2),
  concu = rep(c("ng/mL", "ug/mL"), each = 2),
  conc = c(0, 2, 0, 5),
  time = rep(0:1, 2),
  timeu = "h"
)
d_dose <- data.frame(
  subj = 1,
  dose = 100,
  doseu = "mg",
  time = 0,
  timeu = "h"
)
o_conc <- PKNCA::PKNCAconc(d_conc, conc ~ time | subj / analyte, concu = "concu", timeu = "timeu")
o_dose <- PKNCA::PKNCAdose(d_dose, dose ~ time | subj, doseu = "doseu")
combined_data <- PKNCA_left_join_conc_dose(o_conc, o_dose)

}
