% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCDISC.R
\docType{data}
\name{pptestcd_dict}
\alias{pptestcd_dict}
\title{Export CDISC Data}
\format{
An object of class \code{character} of length 18.
}
\usage{
pptestcd_dict
}
\value{
A list with two data frames:
\describe{
\item{pp}{A data frame containing the PP (Pharmacokinetic Parameters) domain data.}
\item{adpp}{A data frame containing the ADPP (Analysis Dataset for Pharmacokinetic Parameters) domain data.}
}
}
\description{
This function processes the results from a PKNCA and exports them into CDISC compliant datasets.
Attention: All parameters that do no match pptest dataframe will be lost in this pipeline!
}
\details{
Outputs are the following:
\itemize{
\item pknca_result Output from function call \code{pk.nca()} (formatted)
\item pknca_result_raw Output from function call \code{pk.nca()} (needs to be merged with upper later on but now we avoid merge conflict)
}
}
\keyword{datasets}
