% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_lineplot.R
\name{preprocess_data_for_plot}
\alias{preprocess_data_for_plot}
\title{Prepare Data for PK Lineplot}
\usage{
preprocess_data_for_plot(
  data,
  selected_usubjids,
  selected_analytes,
  selected_pcspec,
  colorby_var,
  time_scale,
  yaxis_scale,
  cycle
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_usubjids, selected_analytes, selected_pcspec, cycle}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to be used for coloring.}

\item{time_scale}{String, either "By Dose Profile" or "Actual Time".}

\item{yaxis_scale}{String, either "Log" or "Linear".}
}
\value{
A processed and filtered data.frame.
}
\description{
Prepare Data for PK Lineplot
}
