% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R, R/plot_data_creation.R
\name{process_data_mean}
\alias{process_data_mean}
\title{Create a Mean PK Line Plot}
\usage{
process_data_mean(
  data,
  selected_analytes,
  selected_pcspec,
  profiles_selected = NULL,
  ylog_scale = FALSE,
  color_by = NULL,
  facet_by = NULL
)

process_data_mean(
  data,
  selected_analytes,
  selected_pcspec,
  profiles_selected = NULL,
  ylog_scale = FALSE,
  color_by = NULL,
  facet_by = NULL
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_analytes, selected_pcspec, profiles_selected}{Inputs for filtering.}

\item{ylog_scale}{Logical, whether to use a logarithmic scale for the y-axis.}

\item{color_by, facet_by}{Optional grouping variables to be included in summary.}
}
\value{
A list with summarised_data and time_col.

A list with summarised_data with Mean, SD, and CIs for the profiles selected,
and time_col- either NFRLT or NRRLT depending on if profiles selected is null.
}
\description{
Create a Mean PK Line Plot

Create a Mean PK Line Plot
}
