% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uploads.R
\name{read_uploaded_file}
\alias{read_uploaded_file}
\title{Read uploaded file
This function reads the uploaded input and attempts to parse
it as either PK data or settings YAML.}
\usage{
read_uploaded_file(path, name)
}
\arguments{
\item{path}{Character string with path to the uploaded file.}

\item{name}{Character string with the name of the uploaded file.}
}
\value{
A list with status, content, name, and type of the uploaded file.
}
\description{
Read uploaded file
This function reads the uploaded input and attempts to parse
it as either PK data or settings YAML.
}
\examples{
# Example of reading a PK data file
data_file <- system.file("shiny/data/example-ADNCA.csv", package = "aNCA")
result_data <- read_uploaded_file(data_file, "example-ADNCA.csv")

}
