% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.R
\name{remove_pp_not_requested}
\alias{remove_pp_not_requested}
\title{Filter Out Parameters Not Requested in PKNCA Results (Pivot Version)}
\usage{
remove_pp_not_requested(pknca_res)
}
\arguments{
\item{pknca_res}{A PKNCA results object containing at least $data$intervals and $result.}
}
\value{
The PKNCA results object with non requested parameters removed from $result.
}
\description{
This function removes parameters from the PKNCA results that were not requested by the user,
using a pivoted approach that also handles bioavailability settings.
}
