% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_to_pknca.R
\name{route_cdisc_to_pknca}
\alias{route_cdisc_to_pknca}
\title{Map CDISC route of administration to PKNCA route}
\usage{
route_cdisc_to_pknca(route)
}
\arguments{
\item{route}{Character vector of CDISC route values}
}
\value{
Character vector of \code{"intravascular"} or
\code{"extravascular"}
}
\description{
Map CDISC route of administration to PKNCA route
}
\keyword{internal}
