% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.R
\name{select_level_grouping_cols}
\alias{select_level_grouping_cols}
\title{Select Relevant Columns for Grouping Levels of a target column}
\usage{
select_level_grouping_cols(df, target_col)
}
\arguments{
\item{df}{A data frame.}

\item{target_col}{A character vector of one or more column names to be reconstructed.}
}
\value{
A data frame containing the target column and the minimal set of columns
necessary to reconstruct the target column(s) levels structure.
}
\description{
This function identifies and returns the minimal set of columns from a data frame
that are necessary to uniquely reconstruct the grouping or stratification defined by
the target column(s). "Relevant columns" are those whose combined values can be used
to generate a factor that matches the levels of the target column(s).
}
\details{
The function removes duplicate, constant, and redundant columns, and then searches
for the smallest combination of columns that can uniquely represent the grouping
structure of the target column(s).
}
