% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_operators.R
\name{set_empty_label}
\alias{set_empty_label}
\title{Set an Empty Label if None Exists}
\usage{
set_empty_label(x)
}
\arguments{
\item{x}{A vector to be checked and potentially assigned an empty "label" attribute.}
}
\value{
The vector with an empty "label" attribute if it did not already have one.
}
\description{
This function sets an empty "label" attribute for a vector if it does not already have one.
}
\examples{
\dontrun{
  # Example usage:
  vec <- c("A", "B", "C")
  vec <- set_empty_label(vec)
  print(attr(vec, "label"))  # Returns ""
}

}
