% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_unit.R
\name{transform_unit}
\alias{transform_unit}
\title{Transform Units}
\usage{
transform_unit(initial_unit, target_unit)
}
\arguments{
\item{initial_unit}{A character string representing the initial unit.}

\item{target_unit}{A character string representing the target unit.}
}
\value{
A numeric value for the conversion factor from the initial to the target unit,
or NA if the units are not convertible.
}
\description{
This function transforms a value from an initial unit to a target unit.
}
\examples{
transform_unit("meter", "kilometer")
transform_unit("sec", "min")
}
