% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{update_slopes_intervals}
\alias{update_slopes_intervals}
\title{Update  a PKNCAdata object for slope calculations}
\usage{
update_slopes_intervals(data, params = NULL, should_impute_c0 = TRUE)
}
\arguments{
\item{data}{A PKNCAdata object containing intervals and dosing data.}

\item{params}{A character vector of parameters to include in slope calculations.}

\item{should_impute_c0}{Logical indicating whether to impute start values for parameters.}
}
\value{
A fully configured \code{PKNCAdata} object.
}
\description{
Update  a PKNCAdata object for slope calculations
}
