% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pk.R
\name{validate_pk}
\alias{validate_pk}
\title{Validates datatable with raw pk data.}
\usage{
validate_pk(pk_data)
}
\arguments{
\item{pk_data}{Object to check.}
}
\value{
Original, unchanged object. If any of the checks do not pass, throws an error.
}
\description{
Validates datatable with raw pk data.
}
\details{
Performs the following checks:
\itemize{
\item If provided variable is of class \code{data.frame}.
\item If number of rows in provided data frame is greater than 0.
}

Throws an error if any of the checks does not pass.
}
