% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_operators.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Labels to a dataset}
\usage{
apply_labels(data)
}
\arguments{
\item{data}{The dataset to which labels will be applied.}
}
\value{
The same dataset with label attributes applied to all columns.
If a column is not present in the labels list, it will be assigned the name of the col.
}
\description{
This function adds "label" attributes to all columns in a dataset
Using the labels_app.csv file which contains common labels for ADPP
and ADPC column names
}
\examples{
\dontrun{
  # Example usage:
  data <- data.frame(USUBJID = c(1, 2, 3), AVAL = c(4, 5, 6))
  data <- apply_labels(data)
  print(attr(data$A, "label"))
}

}
