% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability.R
\name{bioavailability_in_pkncaresult}
\alias{bioavailability_in_pkncaresult}
\title{Add bioavailability to PKNCAresults object
This helper function adds bioavailability (F) data to a PKNCAresults object.
The bioavailabiility is calculated with the calculate_bioavailability function.}
\usage{
bioavailability_in_pkncaresult(res_nca, bioavailability)
}
\arguments{
\item{res_nca}{A list containing non-compartmental analysis (NCA) results,
including concentration and dose data.}

\item{bioavailability}{A data frame with calculated bioavailability values}
}
\value{
A PKNCAresults object with bioavailability data added to the result slot.
}
\description{
Add bioavailability to PKNCAresults object
This helper function adds bioavailability (F) data to a PKNCAresults object.
The bioavailabiility is calculated with the calculate_bioavailability function.
}
