% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability.R
\name{calculate_f}
\alias{calculate_f}
\title{Calculate bioavailability with pivoted output}
\usage{
calculate_f(res_nca, selected_aucs)
}
\arguments{
\item{res_nca}{A list containing non-compartmental analysis (NCA) results,
including concentration and dose data.}

\item{selected_aucs}{A character vector of selected
AUC variables (e.g., \code{c("f_aucinf.obs", "f_auclast")}).}
}
\value{
A pivoted data frame with calculated bioavailability values (\code{f_aucinf}, \code{f_auclast}, etc.)
for individual subjects where IV data is available. If IV data is missing, it estimates
bioavailability using the mean IV AUC for that grouping. Columns are assumed to be in \verb{\%} units.
}
\description{
This function calculates bioavailability (F) based on AUC (Area Under Curve) data
extracted from \code{res_nca}. It computes individual bioavailability
where IV and EX data are available for a subject. If IV data is missing, it estimates
bioavailability using the mean IV values for that grouping. The output is pivoted
such that each row represents all main results summarized for each profile in each
subject. Columns are assumed to be in \verb{\%} units even if not explicitly stated.
}
\details{
\itemize{
\item This function is a wrapper around \code{pknca_calculate_f} that reshapes the output
into a pivoted format.
\item The output includes bioavailability estimates for individual subjects and mean-based
estimates, with columns assumed to be in \verb{\%} units.
}
}
