% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{calculate_ratio_app}
\alias{calculate_ratio_app}
\title{Calculate a Ratio for the App}
\usage{
calculate_ratio_app(
  res,
  parameter,
  test = "(all other levels)",
  reference = "PARAM: Analyte01",
  aggregate_subject = "no",
  adjusting_factor = 1,
  custom_pptestcd = NULL
)
}
\arguments{
\item{res}{A PKNCAresult object.}

\item{parameter}{Character. The PPTESTCD value to use for the calculation.}

\item{test}{Character. The test group (numerator). Default is "(all other levels)".}

\item{reference}{Character. The reference group (denominator).}

\item{aggregate_subject}{Character. Aggregation mode: "yes", "no", or "if-needed".}

\item{adjusting_factor}{Numeric that multiplies the calculated ratio. Default is 1.}

\item{custom_pptestcd}{Optional character. If provided, will be used as the PPTESTCD value.}
}
\value{
A data.frame with the calculated ratios for the specified settings.
}
\description{
Calculate a Ratio for the App
}
