% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_code.R
\name{clean_deparse}
\alias{clean_deparse}
\title{Convert R objects into reproducible R code strings (internal)}
\usage{
clean_deparse(obj, indent = 0)
}
\arguments{
\item{obj}{An R object to convert to a string of R code.}

\item{indent}{Integer indentation level for multi-line outputs.}
}
\value{
A single string containing R code that, when evaluated, will
reconstruct \code{obj} (or a close approximation for complex types).
}
\description{
This internal S3 generic converts common R objects (data frames, lists,
atomic vectors, etc.) into character strings containing R code that will
reconstruct the object. It is used by the app script generator to
serialize \code{session$userData} values into a runnable R script.
}
\keyword{internal}
