% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R
\name{create_indplot}
\alias{create_indplot}
\title{Create an Individual PK Line Plot}
\usage{
create_indplot(
  data,
  selected_usubjids,
  selected_analytes,
  selected_pcspec,
  colorby_var = "USUBJID",
  ylog_scale = FALSE,
  profiles_selected = NULL,
  facet_by = NULL,
  show_threshold = FALSE,
  threshold_value = 0,
  show_dose = FALSE,
  dose_data = NULL,
  palette = NULL
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_usubjids, selected_analytes, selected_pcspec, profiles_selected}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to be used for coloring.}

\item{ylog_scale}{Logical, whether to use a logarithmic scale for the y-axis.}

\item{facet_by}{Character vector of column names to facet by.}

\item{show_threshold}{Logical, whether to show threshold line.}

\item{threshold_value}{Numeric, value for the threshold line.}

\item{show_dose}{Logical, whether to show dosing indicators.}

\item{dose_data}{The raw data frame (or a derivative) for calculating dose times.}

\item{palette}{A named color palette.}
}
\value{
A \code{ggplot} object.
}
\description{
Combines data processing and plotting for individual concentration-time profiles.
}
