% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R
\name{create_meanplot}
\alias{create_meanplot}
\title{Create a Mean PK Line Plot}
\usage{
create_meanplot(
  data,
  selected_analytes,
  selected_pcspec,
  cycle = NULL,
  colorby_var,
  facet_by = NULL,
  yaxis_scale = "lin",
  time_scale = "All Time",
  show_threshold = FALSE,
  threshold_value = 0,
  show_dose = FALSE,
  palette = NULL,
  dose_data = NULL,
  show_sd_min = FALSE,
  show_sd_max = TRUE,
  show_ci = FALSE
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_analytes, selected_pcspec, cycle}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to group and color by.}

\item{facet_by}{The variable(s) to facet by.}

\item{yaxis_scale}{String, either "log" or "lin".}

\item{time_scale}{String, either "All Time" or "By Dose Profile".}

\item{show_threshold}{Logical, whether to show threshold line.}

\item{threshold_value}{Numeric, value for the threshold line.}

\item{show_dose}{Logical, whether to show dosing indicators.
@param palette A named color palette.}

\item{dose_data}{The raw data frame (or a derivative) for calculating dose times.}

\item{show_sd_min, show_sd_max, show_ci}{Logicals for error bars.}
}
\value{
A \code{ggplot} object.
}
\description{
Combines data processing and plotting for mean concentration-time profiles.
}
