% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_study_types.R
\name{detect_study_types}
\alias{detect_study_types}
\title{Detect study types}
\usage{
detect_study_types(data, route_column, volume_column = NULL)
}
\arguments{
\item{data}{The dataset containing the study types to be identified.
Assumed to be the output from aNCA formatted concentration data.
Must contain the columns \code{STUDYID}, \code{DRUG}, \code{USUBJID}, \code{PCSPEC},
\code{DOSNOA} (created in \code{format_pkncaconc_data}) and the specified route column.}

\item{route_column}{A character string specifying the
column name for the route of administration.}

\item{volume_column}{A character string specifying the
column name for the volume of a sample. Extravascular
samples must be written as \code{extravascular}.
Can be set to \code{NULL} if not applicable.}
}
\value{
A data frame with the unique combinations of study types,
including the grouping columns and the identified type.
}
\description{
This function detects the type of study based on the provided data.
}
\details{
The function identifies a possible five different types of studies
based on grouping by \code{STUDYID}, \code{DRUG}, \code{USUBJID}, \code{PCSPEC}, and the route column.
The study types are determined as follows:
\itemize{
\item "Excretion Data": If the volume column for a group is greater than 0.
\item "Single Extravascular Dose": If there is only one dose and TAU is NA,
and the route is extravascular.
\item "Single IV Dose": If there is only one dose and TAU is NA,
and the route is not extravascular.
\item "Multiple Extravascular Doses": If there are multiple doses (or TAU is available)
and the route is extravascular.
\item "Multiple IV Doses": If there are multiple doses (or TAU is available)
and the route is not extravascular.
\item If none of these conditions are met, the type is marker as "Unknown".
}
}
\examples{
sample_data <- data.frame(
  STUDYID = "STUDY001",
  DRUG = "Drug",
  USUBJID = c(
    # 1. Single IV Dose subject
    "Subj-01", "Subj-01",
    # 2. Multiple Extravascular Doses subject (identified by TAU)
    "Subj-02", "Subj-02",
    # 3. Excretion Data subject (identified by positive volume)
    "Subj-03", "Subj-03"
  ),
  PCSPEC = "PLASMA",
  DOSNOA = c(
    1, 1,        # Single dose
    1, 1,        # Appears as single dose...
    1, 1         # Single dose
  ),
  ROUTE = c(
    "INTRAVENOUS", "INTRAVENOUS",
    "extravascular", "extravascular",
    "INTRAVENOUS", "INTRAVENOUS"
  ),
  SAMPLE_VOLUME = c(
    NA, 0,
    NA, 0,
    10, 12       # Positive volume indicates excretion
  ),
  TAU = c(
    NA, NA,
    24, 24,      # ...but TAU indicates a multiple-dose regimen
    NA, NA
  )
)

study_summary <- detect_study_types(
  data = sample_data,
  route_column = "ROUTE",
  volume_column = "SAMPLE_VOLUME"
)

}
