% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slope_selector.R
\name{.filter_slopes}
\alias{.filter_slopes}
\title{Filter dataset based on slope selections and exclusions}
\usage{
.filter_slopes(data, slopes, profiles)
}
\arguments{
\item{data}{Data to filter. Must be PKNCAdata list, containing the conc element with
PKNCAconc list and appropriate data frame included under data.}

\item{slopes}{Data frame with slopes selection rules, must contain the following columns:
TYPE, PATIENT, PROFILE, IXrange, REASON.}

\item{profiles}{List with available profiles for each PATIENT.}
}
\value{
Original dataset, with is.included.hl, is.excluded.hl and exclude_half.life columns
modified in accordance to the provided slope filters.
}
\description{
This function filters main dataset based on provided slope selections an exclusions.
}
