% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdisc.R
\name{export_cdisc}
\alias{export_cdisc}
\title{Export CDISC Data}
\usage{
export_cdisc(res_nca)
}
\arguments{
\item{res_nca}{Object with results of the NCA analysis.}
}
\value{
A list with two data frames:
\describe{
\item{pp}{A data frame containing the PP (Pharmacokinetic Parameters) domain data.}
\item{adpp}{A data frame containing the ADPP (Analysis Dataset for Pharmacokinetic Parameters)
domain data.}
}
}
\description{
This function processes the results from a PKNCA and exports them into CDISC compliant datasets.
Attention: All parameters that do no match pptest dataframe will be lost in this pipeline!
}
\details{
Outputs are the following:
\itemize{
\item pknca_result Output from function call \code{pk.nca()} (formatted)
\item pknca_result_raw Output from function call \code{pk.nca()} (needs to be merged with upper later
on but now we avoid merge conflict)
}
}
