% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_breaks.R
\name{filter_breaks}
\alias{filter_breaks}
\title{Filter Breaks for X-Axis}
\usage{
filter_breaks(breaks = NA, plot = plot, min_cm_distance = 0.5, axis = "x")
}
\arguments{
\item{breaks}{A numeric vector of x-axis breaks.}

\item{plot}{A ggplot object used to extract plot dimensions and scales.}

\item{min_cm_distance}{A numeric of the minimum distance between breaks.}

\item{axis}{Axis to filter on, either "x" or "y".}
}
\value{
A numeric vector of filtered x-axis breaks.
}
\description{
Filters X-axis for consecutive breaks with at least the specified distance.
}
\author{
Gerardo Rodriguez
}
