% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_pkncaconc_data}
\alias{format_pkncaconc_data}
\title{Create PK Concentration Dataset}
\usage{
format_pkncaconc_data(ADNCA, group_columns, time_column = "AFRLT")
}
\arguments{
\item{ADNCA}{A data frame containing the ADNCA data.}

\item{analyte}{A character string specifying the analyte of interest.}
}
\value{
A data frame containing the filtered and processed concentration data.
}
\description{
This function creates a pharmacokinetic concentration dataset from the provided ADNCA data.
}
\details{
The function performs the following steps:
\itemize{
\item Creates a 'groups' column by concatenating 'USUBJID' and 'DOSNO'.
\item Arranges and groups the data by groups_column.
\item Adds an index column ('IX') 1:n within each group of length n.
}
}
\examples{
\dontrun{
  # Example usage:
  conc_data <- format_pkncaconc_data(ADNCA, "analyte_name", "profile_type")
}

}
