% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_pkncaconc_data}
\alias{format_pkncaconc_data}
\title{Create PK Concentration Dataset}
\usage{
format_pkncaconc_data(
  ADNCA,
  group_columns,
  time_column = "AFRLT",
  rrlt_column = "ARRLT",
  route_column = "ROUTE"
)
}
\arguments{
\item{ADNCA}{A data frame containing the ADNCA data.}

\item{group_columns}{A character vector specifying the columns to group by.}

\item{time_column}{A character string specifying the time column.}

\item{rrlt_column}{A character string specifying the time since last dose column.}

\item{route_column}{A character string specifying the route column.}
}
\value{
A data frame containing the filtered and processed concentration data.
}
\description{
This function creates a pharmacokinetic concentration dataset from the provided ADNCA data.
}
\details{
The function performs the following steps:
\itemize{
\item Groups the data by groups_column.
\item Calculates \code{TIME_DOSE}as the time of dose reference by the PK sample
\item Creates \code{DOSNOA} variable, sequential numbers based on time of dose
\item Adds a 'std_route' column taking values "intravascular" or "extravascular".
\item Arranges the data by group_columns.
}
}
\examples{
\dontrun{
  # Example usage:
  conc_data <- format_pkncaconc_data(ADNCA,
                                     group_columns,
                                     "AFRLT",
                                     "ROUTE")
}

}
