% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_pkncadose_data}
\alias{format_pkncadose_data}
\title{Create PK Dose Dataset}
\usage{
format_pkncadose_data(
  pkncaconc_data,
  group_columns,
  dosno_column = NULL,
  time_column = "AFRLT",
  since_lastdose_time_column = "ARRLT"
)
}
\arguments{
\item{ADNCA_conc}{A data frame containing the concentration data.}
}
\value{
A data frame containing the dose data.
}
\description{
This function creates a pharmacokinetic dose dataset from the provided concentration data.
}
\details{
The function performs the following steps:
\itemize{
\item Arranges and groups the data by group_columns
\item Derives the dose time as: time_column - since_lastdose_time_column
\item Selects the first row within each group (arranged by since_lastdose_time_column)
}
}
