% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{g_lineplot}
\alias{g_lineplot}
\title{Generate a Unified Line Plot for PK Data}
\usage{
g_lineplot(
  data,
  x_var,
  y_var,
  color_by,
  facet_by = NULL,
  group_by = NULL,
  ylog_scale = FALSE,
  threshold_value = NULL,
  dose_data = NULL,
  palette = NULL,
  ci = FALSE,
  tooltip_vars = NULL,
  labels_df = NULL,
  vline_var = NULL
)
}
\arguments{
\item{data}{A data.frame containing the data to be plotted. This should be
pre-processed by either \code{process_data_individual} or \code{process_data_mean}.}

\item{x_var}{A character string specifying the column name for the x-axis.}

\item{y_var}{A character string specifying the column name for the y-axis.}

\item{color_by}{A character vector specifying the column(s) from the original
dataset that are used to determine the color of the lines and points.}

\item{facet_by}{A character vector of column names to facet the plot by.
Default is \code{NULL} for no faceting.}

\item{group_by}{A character string specifying the column name used to group
the lines. Default is NULL for no grouping.}

\item{ylog_scale}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether to use
a logarithmic scale for the y-axis.}

\item{threshold_value}{A numeric value for the y-intercept of the threshold line.
Only used if \code{show_threshold} is \code{TRUE}.}

\item{dose_data}{An optional data.frame containing dosing information.
If not \code{NULL}, dose lines will be added.
Default is \code{NULL}. The data frame must contain the faceting variables
(if any are used) and a \code{TIME_DOSE} column for the x-intercepts.}

\item{palette}{An optional named character vector of colors for the plot. The names should
correspond to the levels of the \code{color_var} in the \code{data}. This is typically
generated by \code{get_persistent_palette}.}

\item{ci}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots a 95\%
confidence interval ribbon. (Only for mean plots).}

\item{tooltip_vars}{Character vector of column names to include in the tooltip.}

\item{labels_df}{A data.frame for variable label lookups.}

\item{sd_min}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots the lower
standard deviation error bars. (Only for mean plots).}

\item{sd_max}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots the upper
standard deviation error bars. (Only for mean plots).}
}
\value{
A \code{ggplot} object representing the line plot.
}
\description{
This function creates a ggplot2 line plot for pharmacokinetic (PK) data.
It is designed to be flexible and can generate plots for both individual
concentration-time profiles and mean concentration-time profiles. The function
supports various customizations including log scales, faceting, error bars (for mean plots),
and threshold lines.
}
\section{Plot Types}{

The function's behavior changes based on the arguments provided, for two main plot types:
\itemize{
\item \strong{Individual Plots}:
Set \code{y_var = "AVAL"}
The input \code{data} should be from \code{process_data_individual}.
\item \strong{Mean Plots}:
Set \code{y_var = "Mean"}
The input \code{data} should be from \code{process_data_mean}. The arguments \code{sd_min},
\code{sd_max}, and \code{ci} are only applicable to this plot type.
}
}

\examples{
library(dplyr)
ind_data <- expand.grid(
  time_var = c(0, 1, 2, 4, 8, 12),
  USUBJID = c("Subject1", "Subject2")
) \%>\%
  mutate(
    AVAL = ifelse(USUBJID == "Subject1", 50, 80) * exp(-0.5 * time_var) + rnorm(n(), 0, 1),
    PARAM = "Analyte1",
    DOSEA = "Dose 1",
    RRLTU = "hours",
    AVALU = "ng/mL"
  )

 p <- g_lineplot(
   data = ind_data,
   x_var = "time_var",
   y_var = "AVAL",
   color_by = "USUBJID"
   )
print(p)
}
