% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{g_lineplot}
\alias{g_lineplot}
\title{General Line Plot Function}
\usage{
g_lineplot(data, x_var, y_var, group_var, ...)
}
\arguments{
\item{data}{A processed data frame.}

\item{x_var, y_var, group_var}{Strings specifying the variables for aesthetics.}

\item{...}{Additional plotting parameters passed as a list.}
}
\value{
A ggplot object.
}
\description{
Creates a ggplot line plot for either individual or mean PK data.
}
