% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{g_lineplot}
\alias{g_lineplot}
\title{Generate a Unified Line Plot for PK Data}
\usage{
g_lineplot(
  data,
  x_var,
  y_var,
  group_var,
  colorby_var,
  facet_by = NULL,
  yaxis_scale = "lin",
  show_threshold = FALSE,
  threshold_value = 0,
  show_dose = FALSE,
  dose_data = NULL,
  palette = NULL,
  show_sd_min = FALSE,
  show_sd_max = FALSE,
  show_ci = FALSE
)
}
\arguments{
\item{data}{A data.frame containing the data to be plotted. This should be
pre-processed by either \code{process_data_individual} or \code{process_data_mean}.}

\item{x_var}{A character string specifying the column name for the x-axis.}

\item{y_var}{A character string specifying the column name for the y-axis.}

\item{group_var}{A character string specifying the column name used to group
the lines, ensuring \code{geom_line} connects the correct points.}

\item{colorby_var}{A character vector specifying the column(s) from the original
dataset that are used to determine the color of the lines and points.}

\item{facet_by}{A character vector of column names to facet the plot by.
Default is \code{NULL} for no faceting.}

\item{yaxis_scale}{A character string, either \code{"Lin"} (default) or \code{"Log"}, to set the
y-axis scale.}

\item{show_threshold}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether to
display a horizontal threshold line.}

\item{threshold_value}{A numeric value for the y-intercept of the threshold line.
Only used if \code{show_threshold} is \code{TRUE}.}

\item{show_dose}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether to show
vertical lines at dosing times.}

\item{dose_data}{An optional data.frame containing dosing information.
It is required if \code{show_dose} is \code{TRUE}. The data frame must contain the faceting variables
(if any are used) and a \code{TIME_DOSE} column for the x-intercepts.}

\item{palette}{An optional named character vector of colors for the plot. The names should
correspond to the levels of the \code{color_var} in the \code{data}. This is typically
generated by \code{get_persistent_palette}.}

\item{show_sd_min}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots the lower
standard deviation error bars. (Only for mean plots).}

\item{show_sd_max}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots the upper
standard deviation error bars. (Only for mean plots).}

\item{show_ci}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, plots a 95\%
confidence interval ribbon. (Only for mean plots).}
}
\value{
A \code{ggplot} object representing the line plot.
}
\description{
This function creates a ggplot2 line plot for pharmacokinetic (PK) data.
It is designed to be flexible and can generate plots for both individual
concentration-time profiles and mean concentration-time profiles. The function
supports various customisations including log scales, faceting, error bars (for mean plots),
and threshold lines.
}
\section{Plot Types}{

The function's behavior changes based on the arguments provided, for two main plot types:
\itemize{
\item \strong{Individual Plots}:
Set \code{x_var = "time_var"}, \code{y_var = "AVAL"} and \code{group_var = "USUBJID"}.
The input \code{data} should be from \code{process_data_individual}.
\item \strong{Mean Plots}:
Set \code{x_var = "time_var"}, \code{y_var = "Mean"}, and \code{group_var = "color_var"}.
The input \code{data} should be from \code{process_data_mean}. The arguments \code{show_sd_min},
\code{show_sd_max}, and \code{show_ci} are only applicable to this plot type.
}
}

