% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_lineplot.R
\name{general_lineplot}
\alias{general_lineplot}
\title{Generate a General Line Plot for ADNCA Dataset}
\usage{
general_lineplot(
  data,
  selected_analytes,
  selected_usubjids,
  colorby_var,
  time_scale,
  xaxis_scale,
  cycle = NULL
)
}
\arguments{
\item{data}{A data frame containing the ADNCA dataset.}

\item{selected_analytes}{A character vector of selected analytes to be included in the plot.}

\item{selected_usubjids}{A character vector of selected unique subject identifiers (USUBJIDs)
to be included in the plot.}

\item{colorby_var}{A character string specifying the variable by which to color
the lines in the plot.}

\item{time_scale}{A character string specifying the time scale.
Options are "By Cycle" or other values.}

\item{xaxis_scale}{A character string specifying the x-axis scale.
Options are "Log" or other values.}

\item{cycle}{A character string or numeric value specifying the cycle to filter by
when \code{time_scale} is "By Cycle". Default is NULL.}
}
\value{
A ggplot object representing the line plot of pharmacokinetic concentration over time.
}
\description{
This function generates a line plot for an ADNCA dataset based on user-selected analytes,
subjects, and other parameters. The plot can be customized to display data on a linear or
logarithmic scale and can be filtered by cycle.
}
\details{
The function performs the following steps:a
\itemize{
\item Filters the data based on the selected analytes and subjects.
\item Selects relevant columns and removes rows with missing concentration values.
\item Converts 'USUBJID', 'DOSNO', and 'DOSEA' to factors.
\item Filters the data by cycle if \code{time_scale} is "By Cycle".
\item Adjusts concentration values for logarithmic scale if \code{xaxis_scale} is "Log".
\item Generates a line plot using the \code{g_ipp} function with the specified parameters.
\item Adjusts the y-axis to logarithmic scale if \code{xaxis_scale} is "Log".
}
}
\examples{
\dontrun{
  # Example usage:
  plot <- general_lineplot(data = adnca_data,
                           selected_analytes = c("Analyte1", "Analyte2"),
                           selected_usubjids = c("Subject1", "Subject2"),
                           colorby_var = "DOSNO",
                           time_scale = "By Cycle",
                           xaxis_scale = "Log",
                           cycle = "1")
  print(plot)
}

}
