% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_halflife_plots.R
\name{get_halflife_plots}
\alias{get_halflife_plots}
\title{Create a Plotly Half-life Plot}
\usage{
get_halflife_plots(pknca_data, add_annotations = TRUE)
}
\arguments{
\item{pknca_data}{PKNCA data object}

\item{add_annotations}{Logical, whether to add the subtitle annotation}
}
\value{
A list with plotly objects and data
}
\description{
Generates a plotly plot for NCA half-life visualization, with a fit line and scatter points.
}
