% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_code.R
\name{get_settings_code}
\alias{get_settings_code}
\title{Generate a session script from settings and mapping files}
\usage{
get_settings_code(
  settings_file_path,
  data_path,
  output_path = "settings_code.R",
  template_path = NULL,
  mapping = default_mapping,
  ratio_table = data.frame()
)
}
\arguments{
\item{settings_file_path}{Path to the RDS file containing the settings list.}

\item{data_path}{Path to the data file to be referenced in the script.}

\item{output_path}{Path to write the resulting script file.}

\item{template_path}{Path to the R script template file. By default, uses the one
installed from your aNCA package version.}

\item{mapping}{Named list mapping variable names (default: \code{default_mapping}).}

\item{ratio_table}{Data frame containing ratio definitions (default: empty data frame).}
}
\value{
Invisibly returns the output_path.
}
\description{
This function reads a settings RDS file and data path, and generates an R script
that can reproduce the session using a template.
}
