% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_pkconc.R
\name{l_pkconc}
\alias{l_pkconc}
\title{Create PK Concentration Listing}
\usage{
l_pkconc(
  adpc = read.csv("inst/shiny/data/DummyRO_ADNCA.csv"),
  listgroup_vars = c("PARAM", "PCSPEC", "ROUTE"),
  grouping_vars = c("TRT01A", "USUBJID", "AVISIT"),
  displaying_vars = c("NFRLT", "AFRLT", "AVAL"),
  formatting_vars_table = NULL,
  subtitle_lists = NULL,
  footnote_table = "*: Patients excluded from the summary table and mean plots"
)
}
\arguments{
\item{adpc}{A data frame containing the PK concentration data.}

\item{listgroup_vars}{A character vector specifying the variables to group by in the listing.}

\item{grouping_vars}{A character vector specifying the grouping variables for the listing.}

\item{displaying_vars}{A character vector specifying the variables to display in the listing.}

\item{formatting_vars_table}{A data frame specifying the formatting options for the variables.}

\item{subtitle_lists}{A character string specifying the subtitle of the listing table.}

\item{footnote_table}{A character string specifying the footnote of the listing table.}
}
\value{
A list of listings, each corresponding to a unique combination of the grouping variables.
}
\description{
This function creates a listing of pharmacokinetic (PK) concentration data by treatment group,
subject, and nominal time.
}
\details{
The function performs the following steps:
\itemize{
\item Groups the data based on the specified grouping variables.
\item Formats the 0 values as defined by the formatting table.
\item Creates a listing for each unique combination of the grouping variables.
\item Adds units to the variable labels in the listing.
}

The \code{formatting_vars_table} should be a data frame with the following columns:
\itemize{
\item \code{var_name}: The name of the variable.
\item \code{Label}: The label for the variable.
\item \code{na_str}: The string to use for NA values.
\item \code{cero_str}: The string to use for 0 values.
\item \code{align}: The alignment for the variable (e.g., "center").
\item \code{format_fun}: The formatting function to use ("round" or "signif").
\item \code{digits}: The number of digits to use for numeric formatting.
\item \code{unit}: The unit column for the variable (if applicable).
}
}
\examples{
\dontrun{
  # Example usage:
  listings <- l_pkconc(adpc, listgroup_vars, grouping_vars, displaying_vars)
}

}
\author{
Gerardo Rodriguez
}
