% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_variables}
\alias{metadata_variables}
\title{Pharmacokinetic Parameters (ADPC) and Pharmacodynamic Parameters (PP)}
\format{
A data frame with 361 rows and 14 variables:
\describe{
\item{Dataset}{Character. Indicates the dataset the variable belongs to (PP, ADPC, ADPP).}
\item{Order}{Numeric. The sequential order of the variable within its domain. Established based on Role, Core and Variable}
\item{Variable}{Character. The short name of the variable.}
\item{Label}{Character. A descriptive label for the variable.}
\item{Type}{Character. The data type of the variable (e.g., Char, Num, text, integer, float, dateTime).}
\item{Role}{Character. The CDISC role of the variable (e.g., Identifier, Topic, Timing, Result Qualifier, Record Qualifier, Grouping Qualifier, Synonym Qualifier, Variable Qualifier).}
\item{Core}{Character. Indicates the core status of the variable (Req = Required, Perm = Permissible, Exp = Expected, Cond = Conditional).}
\item{Roche_Specific}{Character. Indicates if the variable is Roche-specific.}
\item{is.core}{Logical. TRUE if the variable is a core variable, otherwise FALSE.}
\item{Length}{Numeric. The maximum length of the variable.}
\item{Controlled_Terms}{Character. Reference to controlled terminology (e.g., C85839, C66731).}
\item{is.used}{Logical. TRUE if the variable is used, otherwise FALSE.}
\item{Values}{Character. Possible values for the variable (if applicable).}
\item{is.mapped}{Logical. TRUE if the variable is mapped, otherwise FALSE.}
}
}
\source{
Used for PP and ADPP mapping rules and checks in the export_cdisc function
}
\usage{
metadata_variables
}
\description{
A dataset containing pharmacokinetic variable specifications.
}
\keyword{datasets}
