% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_calculations.R
\name{multiple_matrix_ratios}
\alias{multiple_matrix_ratios}
\title{Calculate Matrix Ratios
This function calculates the  ratios for a given data set,
based on the shared time points for each matrix concentration sample.
The user can input multiple tissues for which ratios should be calculated.}
\usage{
multiple_matrix_ratios(
  data,
  matrix_col,
  conc_col,
  units_col,
  groups = c("TIME", "USUBJID"),
  spec1,
  spec2
)
}
\arguments{
\item{data}{A data frame containing the concentration data.}

\item{matrix_col}{A character string specifying the column name for the matrix type.}

\item{conc_col}{A character string specifying the column name for the concentration data.}

\item{units_col}{A character string specifying the column name for the units.}

\item{groups}{A character vector of grouping variables to use for the analysis.
Must include TIME, USUBJID, and optionally, other grouping variables.}

\item{spec1}{A character string specifying the value for
the first specimen type(s) in the matrix_col.}

\item{spec2}{A character string specifying the value for
the second specimen type(s) in the matrix_col.}
}
\value{
A data frame containing the ratios.
}
\description{
The ratios are calculated as specimen1 / specimen 2.
}
