% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkconc_ind.R
\name{pkcg01}
\alias{pkcg01}
\title{Generate PK Concentration-Time Profile Plots}
\usage{
pkcg01(
  adpc = data(),
  xvar = "AFRLT",
  yvar = "AVAL",
  xvar_unit = "RRLTU",
  yvar_unit = "AVALU",
  color_var = NULL,
  color_var_label = NULL,
  xbreaks_var = "NFRLT",
  xbreaks_mindist = 0.5,
  xmin = NA,
  xmax = NA,
  ymin = NA,
  ymax = NA,
  xlab = paste0(xvar, " [", unique(adpc[[xvar_unit]]), "]"),
  ylab = paste0(yvar, " [", unique(adpc[[yvar_unit]]), "]"),
  footnote = NULL,
  plotgroup_vars = c("ROUTE", "PCSPEC", "PARAM", "USUBJID"),
  plotgroup_names = c("Route", "Specimen", "Analyte", "Subject ID"),
  scale = c("LIN", "LOG", "SBS")[1],
  studyid = "STUDYID",
  trt_var = "TRT01A"
)
}
\arguments{
\item{adpc}{A data frame containing the data.}

\item{xvar}{A character string of the variable name for the x-axis.}

\item{yvar}{A character string of the variable name for the y-axis.}

\item{xvar_unit}{A character string of the unit for the x-axis variable.}

\item{yvar_unit}{A character string of the unit for the y-axis variable.}

\item{color_var}{A character string of the variable name for the color.}

\item{color_var_label}{A character string of the color label.}

\item{xbreaks_var}{A character string of the x-axis breaks.}

\item{xmin}{A numeric value specifying the minimum x-axis limit.}

\item{xmax}{A numeric value specifying the maximum x-axis limit.}

\item{ymin}{A numeric value for the minimum y-axis limit.}

\item{ymax}{A numeric value for the maximum y-axis limit.}

\item{xlab}{Character for x-axis label. Defaults: \code{xvar} label & \code{xvar_unit}.}

\item{ylab}{Character for y-axis label. Defaults: \code{yvar} label & \code{yvar_unit}.}

\item{footnote}{A character string of a manual footnote for the plot.}

\item{plotgroup_vars}{A character vector of the variables to group data.}

\item{plotgroup_names}{A character vector of the grouping variable names.}

\item{studyid}{A character string specifying the study ID variable.}

\item{trt_var}{A character string specifying the treatment variable.}

\item{options}{A list of additional options (e.g., display scale).}
}
\value{
A list of ggplot objects for each unique group.
}
\description{
This function generates a list of ggplots for PK concentration-time profiles.
}
\author{
Gerardo Rodriguez
}
