% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_pkcg.R
\name{pkcg02}
\alias{pkcg02}
\title{Generate PK Concentration-Time Profile Plot by Cohort}
\usage{
pkcg02(
  adpc = data(),
  xvar = "AFRLT",
  yvar = "AVAL",
  xvar_unit = "RRLTU",
  yvar_unit = "AVALU",
  color_var = NULL,
  color_var_label = NULL,
  xbreaks_var = "NFRLT",
  xbreaks_mindist = 0.5,
  xmin = NA,
  xmax = NA,
  ymin = NA,
  ymax = NA,
  xlab = paste0(xvar, " [", unique(adpc[[xvar_unit]])[1], "]"),
  ylab = paste0(yvar, " [", unique(adpc[[yvar_unit]])[1], "]"),
  title = NULL,
  subtitle = NULL,
  footnote = NULL,
  plotgroup_vars = c("ROUTE", "PCSPEC", "PARAM", "TRT01A"),
  plotgroup_names = list(ROUTE = "Route", PCSPEC = "Specimen", PARAM = "Analyte", TRT01A
    = "Treatment"),
  scale = c("LIN", "LOG", "SBS")[1],
  studyid = "STUDYID",
  trt_var = "TRT01A"
)
}
\description{
Generate PK Concentration-Time Profile Plot by Cohort
}
\author{
Kezia Kobana
magic numbers for footnote position and margin, work in app up to 4 lines
NOTE: might require some fine tuning down the line, looks fine now
}
