% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_calculate_nca.R
\name{PKNCA_calculate_nca}
\alias{PKNCA_calculate_nca}
\title{Calculates results for PKNCA analysis.}
\usage{
PKNCA_calculate_nca(pknca_data)
}
\arguments{
\item{pknca_data}{Data object created using PKNCA::PKNCAdata() function.}
}
\value{
Results object with... TODO
}
\description{
Calculates results for PKNCA analysis.
}
\details{
TODO
}
\examples{
TODO

}
