% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA_create_data_object.R
\name{PKNCA_create_data_object}
\alias{PKNCA_create_data_object}
\title{Creates a \code{PKNCA::PKNCAdata} object.}
\usage{
PKNCA_create_data_object(adnca_data)
}
\arguments{
\item{adnca_data}{Data table containing ADNCA data.}
}
\value{
\code{PKNCAdata} object with... TODO
}
\description{
Creates a \code{PKNCA::PKNCAdata} object.
}
\details{
TODO
}
\examples{
TODO

}
