% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.R
\name{PKNCA_create_data_object}
\alias{PKNCA_create_data_object}
\title{Creates a \code{PKNCA::PKNCAdata} object.}
\usage{
PKNCA_create_data_object(adnca_data)
}
\arguments{
\item{adnca_data}{Data table containing ADNCA data.}
}
\value{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
\code{PKNCAdata} object with concentration, doses, and units based on ADNCA data.
\code{PKNCAdata} object with concentration, doses, and units based on ADNCA data.
=======
\code{PKNCAdata} object with... TODO
>>>>>>> fdc64a6 (refactor: moved some PKNCA-related logic code to functions)
=======
\code{PKNCAdata} object with concentration, doses, and units based on ADNCA data.
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)
}
\description{
Creates a \code{PKNCA::PKNCAdata} object.
}
\details{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 2237461 (docs: add documentation for PKNCA_create_daat_object)
=======
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)
This function creates a standard PKNCAdata object from ADNCA data.
It requires the following columns in the ADNCA data:
\itemize{
\item STUDYID: Study identifier.
\item PCSPEC: Matrix.
\item ROUTE: Route of administration.
\item DRUG: Drug identifier.
\item USUBJID: Unique subject identifier.
\item DOSNO: Dose profile number.
\item ANALYTE: Analyte.
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
\item AVAL: Analysis value.
\item AVALU: AVAL unit.
=======
\item AVAL: Concentration value.
\item AVALU: Concentration unit.
>>>>>>> 2237461 (docs: add documentation for PKNCA_create_daat_object)
=======
\item AVAL: Änalysis value.
=======
\item AVAL: Analysis value.
>>>>>>> e19813b (docs: docs)
\item AVALU: AVAL unit.
>>>>>>> fbf4001 (add testing for pkncadata object creation)
=======
\item AVAL: Concentration value.
\item AVALU: Concentration unit.
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)
\item DOSEA: Dose amount.
\item DOSEU: Dose unit.
\item AFRLT: Actual time from first dose.
\item ARRLT: Actual time from reference dose.
\item NFRLT: Nominal time from first dose.
\item ADOSEDUR: Duration of dose.
\item RRLTU: Time unit.
}
\enumerate{
\item Creating pk concentration data using \code{format_pkncaconc_data()}.
\item Creating dosing data using \code{format_pkncadose_data()}.
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
\item Creating \code{PKNCAconc} object using \code{PKNCA::PKNCAconc()}.
=======
\item Creating PKNCAconc object using \code{PKNCA::PKNCAconc()}.
>>>>>>> 2237461 (docs: add documentation for PKNCA_create_daat_object)
=======
\item Creating \code{PKNCAconc} object using \code{PKNCA::PKNCAconc()}.
>>>>>>> e19813b (docs: docs)
=======
\item Creating PKNCAconc object using \code{PKNCA::PKNCAconc()}.
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)
with formula \code{AVAL ~ TIME | STUDYID + PCSPEC + DRUG + USUBJID / ANALYTE}.
\item Creating PKNCAdose object using \code{PKNCA::PKNCAdose()}.
with formula \code{DOSEA ~ TIME | STUDYID + PCSPEC + DRUG + USUBJID}.
\item Creating PKNCAdata object using \code{PKNCA::PKNCAdata()}.
\item Updating units in PKNCAdata object so each analyte has its own unit.
}
<<<<<<< HEAD
<<<<<<< HEAD
}
\examples{
adnca_data <- data.frame(
STUDYID = rep("STUDY001", 6),
PCSPEC = rep("Plasma", 6),
ROUTE = rep("IV", 6),
DRUG = rep("DrugA", 6),
USUBJID = rep("SUBJ001", 6),
DOSNO = rep(1, 6),
ANALYTE = rep("AnalyteA", 6),
AVAL = c(0, 5, 10, 7, 3, 1),
AVALU = rep("ng/mL", 6),
DOSEA = rep(100, 6),
DOSEU = rep("mg", 6),
AFRLT = c(0, 1, 2, 3, 4, 6),
ARRLT = c(0, 1, 2, 3, 4, 6),
NFRLT = c(0, 1, 2, 3, 4, 6),
ADOSEDUR = rep(0.5, 6),
RRLTU = rep("hour", 6)
)
PKNCA_create_data_object(adnca_data)
=======
=======
>>>>>>> fdc64a6 (refactor: moved some PKNCA-related logic code to functions)
TODO
}
\examples{
TODO
<<<<<<< HEAD
>>>>>>> c5a4e95 (refactor: moved some PKNCA-related logic code to functions)
=======
}
\examples{
=======
}
\examples{
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)
adnca_data <- data.frame(
STUDYID = rep("STUDY001", 6),
PCSPEC = rep("Plasma", 6),
ROUTE = rep("IV", 6),
DRUG = rep("DrugA", 6),
USUBJID = rep("SUBJ001", 6),
DOSNO = rep(1, 6),
ANALYTE = rep("AnalyteA", 6),
AVAL = c(0, 5, 10, 7, 3, 1),
AVALU = rep("ng/mL", 6),
DOSEA = rep(100, 6),
DOSEU = rep("mg", 6),
AFRLT = c(0, 1, 2, 3, 4, 6),
ARRLT = c(0, 1, 2, 3, 4, 6),
NFRLT = c(0, 1, 2, 3, 4, 6),
ADOSEDUR = rep(0.5, 6),
RRLTU = rep("hour", 6)
)
PKNCA_create_data_object(adnca_data)
<<<<<<< HEAD
>>>>>>> 2237461 (docs: add documentation for PKNCA_create_daat_object)
=======
>>>>>>> fdc64a6 (refactor: moved some PKNCA-related logic code to functions)
=======
>>>>>>> 090d954 (docs: add documentation for PKNCA_create_daat_object)

}
