% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R
\name{process_data_individual}
\alias{process_data_individual}
\title{Process data for individual lineplot}
\usage{
process_data_individual(
  data,
  selected_usubjids,
  selected_analytes,
  selected_pcspec,
  profiles_selected = NULL
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_usubjids, selected_analytes, selected_pcspec, profiles_selected}{Inputs for filters.
Defaults to metadata_nca_variables.}
}
\value{
A list with processed_data and time_col.
}
\description{
Process data for individual lineplot
}
