% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plotting_data.R
\name{process_data_individual}
\alias{process_data_individual}
\title{Process Data for Individual PK Line Plots}
\usage{
process_data_individual(
  data,
  selected_usubjids,
  selected_analytes,
  selected_pcspec,
  colorby_var,
  time_scale,
  yaxis_scale,
  cycle
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_usubjids, selected_analytes, selected_pcspec, cycle}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to be used for coloring.}

\item{time_scale}{String, either "All Time" or "By Dose Profile".}

\item{yaxis_scale}{String, either "log" or "lin".}
}
\value{
A processed and filtered data.frame ready for plotting.
}
\description{
Filters and prepares raw data for creating individual concentration-time plots.
}
