% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_creation.R
\name{process_data_mean}
\alias{process_data_mean}
\title{Create a Mean PK Line Plot}
\usage{
process_data_mean(
  data,
  selected_analytes,
  selected_pcspec,
  profiles_selected = NULL,
  ylog_scale = FALSE
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_analytes, selected_pcspec, profiles_selected}{Inputs for filtering.}

\item{ylog_scale}{Logical, whether to use a logarithmic scale for the y-axis.}
}
\value{
A list with summarised_data and time_col.
}
\description{
Create a Mean PK Line Plot
}
