% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plotting_data.R
\name{process_data_mean}
\alias{process_data_mean}
\title{Process Data for Mean PK Line Plots}
\usage{
process_data_mean(
  data,
  selected_analytes,
  selected_pcspec,
  cycle,
  colorby_var,
  facetby_var = NULL,
  yaxis_scale,
  time_scale
)
}
\arguments{
\item{data}{Raw data frame.}

\item{selected_analytes, selected_pcspec, cycle}{Inputs for filtering.}

\item{colorby_var}{The variable(s) to group and color by.}

\item{facetby_var}{The variable(s) to facet by.}

\item{yaxis_scale}{String, either "log" or "lin".}

\item{time_scale}{String, either "All Time" or "By Dose Profile"}
}
\value{
A summarized data.frame with mean, SD, and CI values.
}
\description{
Filters and summarizes raw data for creating mean concentration-time plots.
}
