% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pk.R
\name{read_pk}
\alias{read_pk}
\title{Reads PK datasets from various file formats.}
\usage{
read_pk(path)
}
\arguments{
\item{path}{Character string with path to the dataset file.}
}
\value{
A data.frame object with loaded data.
}
\description{
Reads PK datasets from various file formats.
}
\details{
Currently supported file formats include:
\itemize{\item \code{csv}
\item \code{rds}
\item \code{xlsx}
\item \code{sas7bdat}
\item \code{xpt}
\item \code{parquet}}
}
\examples{
df <- read_pk(system.file("shiny/data/example-ADNCA.csv", package = "aNCA"))

}
