% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uploads.R
\name{read_settings}
\alias{read_settings}
\title{Helper Logic to parse and structure settings YAML}
\usage{
read_settings(path, name)
}
\arguments{
\item{path}{Character string with path to the settings YAML file.}

\item{name}{Character string with the name of the settings YAML file.}
}
\value{
A list with parsed settings or NULL if not a valid settings file.
}
\description{
Helper Logic to parse and structure settings YAML
}
