% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{rm_impute_obs_params}
\alias{rm_impute_obs_params}
\title{Apply Imputation for PKNCA Data}
\usage{
rm_impute_obs_params(data, metadata_nca_parameters = metadata_nca_parameters)
}
\arguments{
\item{data}{A PKNCAdata object.}

\item{metadata_nca_parameters}{A data frame mapping PKNCA parameters (\code{PKNCA})
and information on their parameter dependencies ('Depends').}
}
\value{
A PKNCAdata object with imputation rules applied.
}
\description{
Applies imputation rules to a PKNCAdata object, imputing start values and then
selectively removing imputation for parameters that are not dependent on AUC.
}
