% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny app}
\usage{
run_app(datapath = NULL, ...)
}
\arguments{
\item{datapath}{Full path to a single \code{.csv} or \code{.rds} data file.}

\item{...}{Arguments passed to \code{shiny::runApp()}}
}
\description{
Run the Shiny app
}
\details{
If a \code{datapath} is provided, the app will attempt to automatically load the
specified dataset on startup.This is achieved by setting an internal option
(\code{options(aNCA.datapath = datapath)}), which the app then reads.

If \code{datapath} is \code{NULL} (default), the app will launch without pre-loading any
data, and a file must be uploaded manually within the app.
}
\examples{
\dontrun{
  # Create a dummy data file
  temp_data_path <- tempfile(fileext = ".csv")
  write.csv(data.frame(USUBJID = 1, AFRLT = 0:4, AVAL = c(0, 10, 8, 4, 1)),
            file = temp_data_path, row.names = FALSE)

  # Run the app, automatically loading the dummy data
  run_app(datapath = temp_data_path)

  # Run the app without pre-loading data (standard usage)
  run_app()
}

}
