% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny app}
\usage{
run_app()
}
\description{
List of functions imported for the shiny application.
When adding new imports, please keep the alphabetical order, at lest for packages.
}
