% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{update_main_intervals}
\alias{update_main_intervals}
\title{Update an intervals data frame with user-selected parameters by study type}
\usage{
update_main_intervals(
  data,
  parameter_selections,
  study_types_df,
  auc_data,
  impute = TRUE,
  blq_imputation_rule = NULL
)
}
\arguments{
\item{data}{A PKNCAdata object containing intervals and dosing data.}

\item{parameter_selections}{A named list of selected PKNCA parameters by study type.}

\item{study_types_df}{A data frame mapping analysis profiles to their study type.}

\item{auc_data}{A data frame containing partial AUC ranges.}

\item{impute}{Logical indicating whether to impute start values for parameters.}

\item{blq_imputation_rule}{A list defining the Below Limit of Quantification (BLQ)
imputation rule using PKNCA format. The list should either contain three elements named:
\code{first}, \code{middle}, and \code{last} or two elements named \code{before.tmax} and \code{after.tmax}.
Each element can be a numeric value (substituting the BLQ value), or a string such as
\code{"drop"} (ignores the value) or \code{"keep"} (keeps the value as 0). Default is NULL,
which does not specify any BLQ imputation in any interval.}
}
\value{
An updated PKNCAdata object with parameter intervals based on user selections.
}
\description{
Update an intervals data frame with user-selected parameters by study type
}
